% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_pcid}
\alias{add_pcid}
\title{Add PCID to GESS results}
\usage{
add_pcid(gess_tb)
}
\arguments{
\item{gess_tb}{tibble of GESS result, can be accessed by the `result` method 
on the \code{\link{gessResult}} object}
}
\value{
tibble object where the PCID column is added
}
\description{
This function can be used to add the \code{PCID} (PubChem CID) column to the
GESS result table.
}
\examples{
data("lincs_pert_info")
# gess_tb2 <- add_pcid(gess_tb)
}
