% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_custom_db.R
\name{build_custom_db}
\alias{build_custom_db}
\title{build_custom_db}
\usage{
build_custom_db(df, dest_dir)
}
\arguments{
\item{df}{data.frame, represents genome-wide Gene Expression Profiles (GEPs)
of compound or genetic treatments in cells. The GEPs can be log2 fold change,
z-scores \emph{etc.} from differential expression analysis, or gene 
expression intensity values from Affymetrix Chips, or read counts from 
RNA-Seq data. 

Rownames (Gene SYMBOLs) should be included. The colnames should be of 
`(drug)__(cell)__(factor)` format, e.g., `sirolimus__MCF7__trt_cp`.
It can be generalized to any type of treatment, for example, gene knockdown
or overexpression by setting `drug` as gene symbol, `factor` as 'gene_ko' or 
other words, `cell` as treatment cell types. So, one example for 
generalization format could be `P53__MCF7__gene_ko`.}

\item{dest_dir}{directory path to store the reference database, which is 
the hdf5 backed \code{\link{SummarizedExperiment}} object}
}
\value{
hdf5 backed \code{\link{SummarizedExperiment}} object
}
\description{
Build custom reference signature database for GESS
}
\examples{
# Generate a data.frame 
df <- data.frame(sirolimus__MCF7__trt_cp=rnorm(1000),
                 vorinostat__SKB__trt_cp=rnorm(1000))
data(targetList)
rownames(df) = names(targetList)
dest_dir = tempfile("h5_df_")
# The following step takes some time, so do not run as example
# se = build_custom_db(df, dest_dir)
list.files(dest_dir)
unlink(dest_dir, recursive=TRUE)
}
\seealso{
\code{\link{saveHDF5SummarizedExperiment}}
}
