% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{cell_info}
\alias{cell_info}
\title{Cell Type Information}
\format{
A \code{tibble} object with 30 rows and 4 columns.
}
\usage{
cell_info
}
\description{
It contains cell type (tumor or normal), primary site and subtype 
annotations of cells in LINCS database.
}
\examples{
# Load object
data(cell_info)
head(cell_info)
}
\keyword{datasets}
