% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{cell_info2}
\alias{cell_info2}
\title{LINCS 2020 Cell Type Information}
\format{
A \code{tibble} object with 240 rows and 21 columns.
}
\usage{
cell_info2
}
\description{
It contains cell type (tumor or normal), primary site, subtype etc.
annotations of cells in LINCS 2020 database.
}
\examples{
# Load object
data(cell_info2)
head(cell_info2)
}
\keyword{datasets}
