% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_res_summary.R
\name{cell_rank_sum}
\alias{cell_rank_sum}
\title{cell_rank_sum}
\usage{
cell_rank_sum(gessResult)
}
\arguments{
\item{gessResult}{`gessResult` object}
}
\value{
tibble
}
\description{
Summarize ranks of drugs in GESS result across cell types
}
