% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_treat_info.R
\name{cellNtestPlot}
\alias{cellNtestPlot}
\title{Number of Tests in Cell Types}
\usage{
cellNtestPlot(refdb)
}
\arguments{
\item{refdb}{character(1), one of "lincs", "lincs_expr", "cmap" or "cmap_expr"
when using the pre-generated CMAP/LINCS databases or path to the HDF5 file
generated with the \code{\link{build_custom_db}} function. The details is 
shown in the 'refdb' argument of the \code{\link{qSig}} function}
}
\value{
Faceted bar plot
}
\description{
Bar plot of number of perturbations/compounds tested in cell types where 
cell types are grouped by 'primary site'.
}
\examples{
refdb <- system.file("extdata", "sample_db.h5", package="signatureSearch")
cellNtestPlot(refdb)
}
