% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{chembl_moa_list}
\alias{chembl_moa_list}
\title{MOA to Gene Mappings}
\format{
An object of class \code{list} of length 1099.
}
\usage{
chembl_moa_list
}
\description{
It is a list containing MOA terms to gene Entrez id mappings from ChEMBL 
database
}
\examples{
# Load object
data(chembl_moa_list)
head(chembl_moa_list)
}
\keyword{datasets}
