% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{clue_moa_list}
\alias{clue_moa_list}
\title{MOA to Drug Name Mappings}
\format{
An object of class \code{list} of length 701.
}
\usage{
clue_moa_list
}
\description{
It is a list containing MOA terms to drug name mappings obtained from 
Touchstone database at CLUE website (https://clue.io/)
}
\examples{
# Load object
data(clue_moa_list)
head(clue_moa_list)
}
\keyword{datasets}
