% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feaResult.R
\docType{methods}
\name{cnetplot,feaResult-method}
\alias{cnetplot,feaResult-method}
\alias{cnetplot,feaResult,ANY-method}
\title{cnetplot for feaResult}
\usage{
\S4method{cnetplot}{feaResult}(x, showCategory = 5,
  categorySize = "pvalue", foldChange = NULL, fixed = TRUE, ...)
}
\arguments{
\item{x}{feaResult object}

\item{showCategory}{number of enriched terms to display}

\item{categorySize}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{foldChange}{fold change}

\item{fixed}{TRUE or FALSE}
}
\description{
cnetplot for feaResult
}
