% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_fea_res.R
\name{comp_fea_res}
\alias{comp_fea_res}
\title{plot comparison of FEA enrichment results}
\usage{
comp_fea_res(dup_hyperG_res = NULL, mgsea_res = NULL,
  mabs_res = NULL, hyperG_res = NULL, gsea_res = NULL, Ntop = 20,
  type = "GO")
}
\arguments{
\item{dup_hyperG_res}{enrichment result from dup_hyperG method}

\item{mgsea_res}{enrichment result from mGSEA method}

\item{mabs_res}{enrichment result from mabs method}

\item{hyperG_res}{enrichment result from hyperG method}

\item{gsea_res}{enrichment result from GSEA method}

\item{Ntop}{number of top enrichment result from each method used to do comparison. The plot only show the top 20 ranking terms}

\item{type}{"GO" or "KEGG", indicating the type of enriched terms}
}
\value{
igraph object
}
\description{
Plot consistency comparison of enriched terms from FEA methods
}
