% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feaResult.R
\docType{methods}
\name{dotplot,feaResult-method}
\alias{dotplot,feaResult-method}
\alias{dotplot,feaResult,ANY-method}
\title{dotplot for feaResult}
\usage{
\S4method{dotplot}{feaResult}(object, x = "geneRatio",
  colorBy = "p.adjust", showCategory = 10, split = NULL,
  font.size = 12, title = "")
}
\arguments{
\item{object}{an instance of feaResult}

\item{x}{variable for x axis}

\item{colorBy}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{showCategory}{number of category}

\item{split}{separate result by 'category' variable}

\item{font.size}{font size}

\item{title}{plot title}
}
\description{
dotplot for feaResult
}
