% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{drugs}
\alias{drugs}
\title{Drug names used in examples}
\format{An object of class \code{character} of length 10.}
\usage{
drugs
}
\description{
A character vector containing the names of the top 10 drugs in the GESS 
result from the gess_lincs method used in the vignette of signatureSearch.
}
\examples{
# Load drugs object
data(drugs)
drugs
}
\keyword{datasets}
