% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{drugs10}
\alias{drugs10}
\title{Drug Names Used in Examples}
\format{An object of class \code{character} of length 10.}
\usage{
drugs10
}
\description{
A character vector containing the names of the top 10 drugs in the GESS 
result from the \code{\link{gess_lincs}} method used in the vignette of 
signatureSearch.
}
\examples{
# Load drugs object
data(drugs10)
drugs10
}
\keyword{datasets}
