% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enricher_internal.R
\name{enricher_internal}
\alias{enricher_internal}
\title{enrich.internal}
\usage{
enricher_internal(gene, pvalueCutoff, pAdjustMethod = "BH", universe = NULL,
  minGSSize = 10, maxGSSize = 500, qvalueCutoff = 0.2, USER_DATA)
}
\arguments{
\item{gene}{a vector of SYMBOL gene id.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{qvalueCutoff}{cutoff of qvalue}

\item{USER_DATA}{ontology information}
}
\value{
A \code{feaResult} instance.
}
\description{
interal method for enrichment analysis by using the hypergeometric model, which also support query gene set with duplications
}
