% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO2.R
\name{enrichGO2}
\alias{enrichGO2}
\title{GO enrichment analysis via hypergeometric test}
\usage{
enrichGO2(gene, OrgDb, keytype = "SYMBOL", ont = "MF",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe,
  qvalueCutoff = 0.2, minGSSize = 5, maxGSSize = 500, pool = FALSE)
}
\arguments{
\item{gene}{a vector of entrez gene id or gene SYMBOL.}

\item{OrgDb}{OrgDb}

\item{keytype}{keytype of input gene}

\item{ont}{One of "MF", "BP", "CC" or "ALL"}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", 
"bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{qvalueCutoff}{qvalue cutoff}

\item{minGSSize}{minimal size of genes annotated by Ontology term}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pool}{If ont='ALL', whether pool 3 GO sub-ontologies}
}
\value{
A \code{feaResult} instance.
}
\description{
Given a vector of genes, this function will return the enriched GO
categories after FDR control.
The hypergeometric test is adjusted to support gene set with duplications
}
\examples{
# The method supports duplicated elements in 'gene', 
# which should be SYMBOL id for GO enrichment.
gene = c(rep("HDAC1",4), rep("HDAC3",2), "SOX8", "KLK14")
library(org.Hs.eg.db)
data(targetList)
\dontrun{
ego <- enrichGO2(gene = gene, OrgDb=org.Hs.eg.db, ont="MF",
                 universe=names(targetList))
head(ego)
}
}
\seealso{
\code{\link{feaResult-class}}
}
