% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGG.R
\name{enrichKEGG}
\alias{enrichKEGG}
\title{KEGG Enrichment Analysis of a gene set.}
\usage{
enrichKEGG(gene, organism = "hsa", keyType = "kegg",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe, minGSSize = 5,
  maxGSSize = 500, qvalueCutoff = 0.2)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{qvalueCutoff}{qvalue cutoff}
}
\value{
A \code{feaResult} instance.
}
\description{
Given a vector of genes, this function will return the enrichment KEGG
categories with FDR control.
}
