% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichKEGG2.R
\name{enrichKEGG2}
\alias{enrichKEGG2}
\title{KEGG pathways enrichment analysis via hypergeometric test}
\usage{
enrichKEGG2(gene, organism = "hsa", keyType = "kegg",
  pvalueCutoff = 0.05, pAdjustMethod = "BH", universe, minGSSize = 5,
  maxGSSize = 500, qvalueCutoff = 0.2)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{organism}{supported organism listed in 
\url{http://www.genome.jp/kegg/catalog/org_list.html}}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncbi-proteinid' and 'uniprot'}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", 
"BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by 
ontology term for testing.}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{qvalueCutoff}{qvalue cutoff}
}
\value{
A \code{feaResult} instance.
}
\description{
Given a vector of genes, this function will return the enriched KEGG
pathways after FDR control.
The hypergeometric test is adjusted to support gene set with duplications
}
\examples{
# It supports duplicated elements in "gene", which should be entrez id
gene = c(rep("4312",4), rep("8318",2), "991", "10874")
#data(geneList, package="DOSE")
#kk <- enrichKEGG2(gene = gene, universe=names(geneList))
#head(kk)
}
