% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{geneID}
\alias{geneID}
\alias{geneID.feaResult}
\title{geneID generic}
\usage{
\method{geneID}{feaResult}(x)
}
\arguments{
\item{x}{feaResult object}
}
\value{
'geneID' return the 'geneID' column of the FEA result which can be converted to data.frame via 'as.data.frame'
}
\description{
geneID generic
}
\examples{
data(drugs, package="signatureSearch")
dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", type = "GO", ont="MF")
head(geneID(dup_hyperG_res))
}
