% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{geneInCategory}
\alias{geneInCategory}
\alias{geneInCategory.feaResult}
\title{geneInCategory generic}
\usage{
\method{geneInCategory}{feaResult}(x)
}
\arguments{
\item{x}{feaResult}
}
\value{
'geneInCategory' return a list of genes, by spliting the input gene vector to enriched functional categories
}
\description{
geneInCategory generic
}
\examples{
data(drugs, package="signatureSearch")
dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", type = "GO", ont="MF")
head(geneInCategory(dup_hyperG_res))
}
