% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_cmap.R
\name{gess_cmap}
\alias{gess_cmap}
\title{gess_cmap}
\usage{
gess_cmap(qSig, chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot of 'qSig' should be 'CMAP'}

\item{chunk_size}{size of chunk per processing}
}
\value{
gessResult object, "a list of drugs in reference database ranked by their similarity to query signature
}
\description{
CMAP method for GESS
}
\details{
Motivation: gene expression signature search (GESS) using original CMap enrichment and new LINCS algorithms, as well as variants from gCMAP package
against very large expression databases containing tens to hundreds of thousands of profiles with very moderate memory requirements. The original 
CMap and new LINCS enrichment functions are custom implementations according to Lamb et al, 2006 and Subramanian et al, 2017. Most methods 
can be easily parallelized for multiple query signatures. The time performance of the gess_cmap/gess_lincs methods on a single CPU core is about 
4 min for querying with a single signature ~100,000 CMAP/LINCS database signatures for gess_cmap method, while it is ~10 min for gess_lincs method.
}
