% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_gcmap.R
\name{gess_gcmap}
\alias{gess_gcmap}
\title{gCMAP method for GESS}
\usage{
gess_gcmap(qSig, higher, lower, chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot of 'qSig' should be 'gCMAP'}

\item{higher}{The 'higher' threshold. If not 'NULL', genes with a score 
larger than 'higher' will be included in the gene set with sign +1. 
At least one of 'lower' and 'higher' must be specified.}

\item{lower}{The lower threshold. If not 'NULL', genes with a score smaller 
than 'lower' will be included in the gene set with sign -1. 
At least one of 'lower' and 'higher' must be specified.}

\item{chunk_size}{size of chunk per processing}
}
\value{
gessResult object, containing drugs in the reference database
ranked by their similarity to the query signature
}
\description{
It uses query signature to search against the reference database in the 
\code{\link{qSig}} object by gCMAP method, which is adapted from the 
gCMAP package (Sandmann et al., 2014)
}
\details{
The \code{gCMAP} package provides access to related 
but not identical implementations of the original CMAP algorithm proposed by 
Lamb et al., 2006. It uses as query a rank transformed GEP and the reference 
database is composed of DEG sets. This is the opposite situation of the 
original \code{CMAP} method, where the query is a DEG set and the database 
contains rank transformed GEPs.

Description of the score columns in the gess_gcmap tibble result:
\itemize{
    \item effect: Scaled bi-directional enrichment score corresponding to 
    the scaled_score under the CMAP result.
    \item nSet: Number of genes in the reference gene sets after applying
    the higher and lower cutoff.
    \item nFound: Number of genes in the reference gene sets that are 
    present in the query signature.
    \item signed: Whether the gene sets in the reference database have signs, 
    e.g. representing up and down regulated genes when computing scores.
}
}
\examples{
db_path <- system.file("extdata", "sample_db.h5", 
                       package = "signatureSearch")
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_gcmap <- qSig(query=query_mat, gess_method="gCMAP", refdb=db_path)
gcmap <- gess_gcmap(qsig_gcmap, higher=1, lower=-1)
result(gcmap)
}
\references{
Sandmann, T., Kummerfeld, S. K., Gentleman, R., & Bourgon, R. 
(2014). gCMAP: user-friendly connectivity mapping with R. Bioinformatics , 
30(1), 127–128. \url{https://doi.org/10.1093/bioinformatics/btt592}
}
\seealso{
\code{\link{qSig}}, \code{\link{gessResult}}, \code{\link{gess}}
}
