% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{gess_lincs}
\alias{gess_lincs}
\title{LINCS Search Method}
\usage{
gess_lincs(qSig, tau = FALSE, sortby = "NCS", chunk_size = 5000)
}
\arguments{
\item{qSig}{\code{\link{qSig}} object defining the query signature, the GESS
method (should be 'LINCS') and the path to the reference database.}

\item{tau}{TRUE or FALSE, whether to compute the tau score. It is only 
meaningful when the full LINCS database is used since the reference
queries we generated for computing Tau scores are from LINCS database.}

\item{sortby}{rank the GESS result by one of the following scores: 
`WTCS`, `NCS`, `Tau`, `NCSct` or `NA`}

\item{chunk_size}{size of chunk per processing}
}
\value{
\code{\link{gessResult}} object, the result table contains the 
search results for each perturbagen in the reference database ranked by 
their signature similarity to the query.
}
\description{
It uses query signature to search against the reference database defined
in the \code{\link{qSig}} object by LINCS method, which is an implementation
of the method from Subramanian et al, 2017.
}
\details{
Subramanian et al. (2017) introduced a more complex GESS algorithm, 
here referred to as LINCS. While related to CMAP, there are several important
differences among the two approaches. First, LINCS weights the query genes 
based on the corresponding differential expression scores of the GESs in the 
reference database (e.g. LFC or z-scores). Thus, the reference database used 
by LINCS needs to store the actual score values rather than their ranks. 
Another relevant difference is that the LINCS algorithm uses a bi-directional
weighted Kolmogorov-Smirnov enrichment statistic (ES) as similarity metric. 
To the best of our knowledge, the LINCS search functionality in this package
provides the first downloadable standalone software implementation of this 
algorithm.

The LINCS method takes about 1 min on a single CPU core for querying 
with a single signature against 10,000 signatures in the reference database.
}
\section{Column description}{

Description of the score columns in the result table specific for LINCS 
method:
\itemize{
  \item WTCS: Weighted Connectivity Score, a bi-directional Enrichment 
  Score for an up/down query set. If the ES values of an up set and a down 
  set are of different signs, then WTCS is (ESup-ESdown)/2, otherwise, 
  it is 0. WTCS values range from -1 to 1. They are positive or negative 
  for signatures that are positively or inversely related, respectively, 
  and close to zero for signatures that are unrelated.
  \item WTCS_Pval: Nominal p-value of WTCS computed by comparing WTCS 
  against a null distribution of WTCS values obtained from a large number
  of random queries (e.g. 1000).
  \item WTCS_FDR: False discovery rate of WTCS_Pval.
  \item NCS: Normalized Connectivity Score. To make connectivity scores 
  comparable across cell types and perturbation types, 
  the scores are normalized. Given a vector of WTCS 
  values w resulting from a query, the values are normalized within each 
  cell line c and perturbagen type t to obtain (NCS) by dividing the WTCS 
  value with signed mean of the WTCS values within 
  the subset of signatures in reference database corresponding to c and t.
  \item Tau: Enrichment score standardized for a given database. 
  Tau score compares an observed NCS to a large set of NCS values that have 
  been pre-computed for a specific reference database. The query results 
  are scored with Tau as a standardized measure ranging from 100 to -100. 
  A Tau of 90 indicates that only 10% of reference perturbations exhibit 
  stronger connectivity to the query. This way one can make more meaningful 
  comparisons across query results.
  \item TauRefSize: Size of reference perturbations for computing Tau.
  \item NCSct: NCS summarized across cell types. Given a vector of NCS values
  for perturbagen p, relative to query q, across all cell lines c in which p 
  was profiled, a cell-summarized connectivity score is obtained using a 
  maximum quantile statistic. It compares the 67 and 33 quantiles of 
  NCSp,c and retains whichever is of higher absolute magnitude.
}
Description of the other columns are available at the 'result' slot of the
\code{\link{gessResult}} object.
}

\examples{
db_path <- system.file("extdata", "sample_db.h5", 
                       package = "signatureSearch")
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_lincs <- qSig(query = list(upset=upset, downset=downset), 
                   gess_method = "LINCS", refdb = db_path)
lincs <- gess_lincs(qsig_lincs, sortby="NCS", tau=FALSE)
result(lincs)
}
\references{
For detailed description of the LINCS method and scores, 
please refer to: Subramanian, A., Narayan, R., Corsello, S. M., Peck, D. D., 
Natoli, T. E., Lu, X., … Golub, T. R. (2017). A Next Generation 
Connectivity Map: L1000 Platform and the First 1,000,000 Profiles. Cell, 
171(6), 1437–1452.e17. \url{https://doi.org/10.1016/j.cell.2017.10.049}
}
\seealso{
\code{\link{qSig}}, \code{\link{gessResult}}, \code{\link{gess}}
}
