% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_sp.R
\name{gess_sp}
\alias{gess_sp}
\title{gess_sp}
\usage{
gess_sp(qSig, chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot of 'qSig' should be 'SP'}

\item{chunk_size}{size of chunk per processing}
}
\value{
gessResult object
}
\description{
Gene expression signature search with Spearman correlation
}
