% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gessResult-class}
\alias{gessResult-class}
\alias{gessResult}
\title{Class "gessResult"}
\description{
The class stores the result of GESS analysis
}
\section{Slots}{

\describe{
\item{\code{result}}{tibble from GESS analysis, represents a list of drugs 
in the reference database ranked by their signature similarity to the query.

The description of the common tibble columns from different GESS methods:
\itemize{
    \item pert: character, name of perturbagen (e.g. drug) in the reference 
    database
    \item cell: character, acronym of cell type
    \item type: character, perturbation type. In CMAP and LINCS 
    databases, the perturbation types are all compound treatment(trt_cp). 
    Users can build their custom signature database with other types of 
    perturbation, e.g., gene knockdown or overexpression via 
    \code{\link{build_custom_db}} function
    \item trend: character, up or down when reference signature is 
    positively or negatively connected with the query signature, 
    respectively.
    \item N_upset: integer, number of genes in the query up set
    \item N_downset: integer, number of genes in the query down set
    \item t_gn_sym: character, gene symbols of the corresponding drug 
    targets
}}

\item{\code{query}}{query signature}

\item{\code{gess_method}}{method for GESS analysis}

\item{\code{refdb}}{path to the reference database}
}}

\keyword{classes}
