% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feaResult-methods.R
\docType{methods}
\name{get_drugs}
\alias{get_drugs}
\alias{get_drugs,feaResult-method}
\title{Extract Drug Names from feaResult}
\usage{
get_drugs(x)

\S4method{get_drugs}{feaResult}(x)
}
\arguments{
\item{x}{feaResult object}
}
\value{
character vector
}
\description{
The \code{get_drugs} generic extracts the drug names/ids stored
in the \code{drugs} slot of an feaResult object.
}
\examples{
data(drugs)
# dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
#                                   type = "GO", ont="MF")
# get_drugs(dup_hyperG_res)
}
