% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO.R
\name{get_GO_data}
\alias{get_GO_data}
\title{get_GO_data}
\usage{
get_GO_data(OrgDb, ont, keytype)
}
\arguments{
\item{OrgDb}{Organism annotation package, e.g., `org.Hs.eg.db` for human}

\item{ont}{GO ontology}

\item{keytype}{keytype}
}
\value{
environment
}
\description{
get environment containing GO data
}
