% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_treat_info.R
\name{get_treat_info}
\alias{get_treat_info}
\title{Get Treatment Information}
\usage{
get_treat_info(refdb)
}
\arguments{
\item{refdb}{character(1), one of "lincs", "lincs_expr", "cmap" or "cmap_expr"
when using the pre-generated CMAP/LINCS databases or path to the HDF5 file
generated with the \code{\link{build_custom_db}} function. The details is 
shown in the 'refdb' argument of the \code{\link{qSig}} function}
}
\value{
tibble object with 'pert', 'cell', 'pert_type' columns
}
\description{
Get treatment information including perturbation name, cell type and 
perturbation type from the reference database
}
\examples{
refdb <- system.file("extdata", "sample_db.h5", package="signatureSearch")
treat_info <- get_treat_info(refdb)
}
