% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functional_catgory_data.R
\name{getALLEG}
\alias{getALLEG}
\title{getALLEG}
\usage{
getALLEG(organism)
}
\arguments{
\item{organism}{one of "human", "rat", "mouse", "celegans", "yeast", 
"zebrafish", "fly".}
}
\value{
entrez gene ID vector
}
\description{
get all entrez gene ID of a specific organism
}
\author{
Yu Guangchuang
}
