% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functional_catgory_data.R
\name{getDb}
\alias{getDb}
\title{getDb}
\usage{
getDb(organism)
}
\arguments{
\item{organism}{one of supported organism}
}
\value{
annotationDb name
}
\description{
mapping organism name to annotationDb package name
}
\author{
Yu Guangchuang
}
