% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getSig}
\alias{getSig}
\title{Drawn GES}
\usage{
getSig(cmp, cell, refdb)
}
\arguments{
\item{cmp}{character(1), compound name available in refdb}

\item{cell}{character(1), cell type that the compound treated in}

\item{refdb}{character(1), one of "cmap", "cmap_expr", "lincs", or 
"lincs_expr"}
}
\value{
matrix with one column representing GES of the query compound in cell
}
\description{
Drawn from reference database (e.g. LINCS, CMAP) a GES of a compound 
treatment in a cell type
}
\examples{
vor_sig <- getSig("vorinostat", "SKB", refdb="notrun")
}
