% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_internal2.R
\name{GSEA_internal2}
\alias{GSEA_internal2}
\title{GSEA_internal2}
\usage{
GSEA_internal2(geneList, exponent, nPerm, minGSSize, maxGSSize, pvalueCutoff,
  pAdjustMethod, verbose, seed = FALSE, USER_DATA, nproc = 1,
  by = "fgsea")
}
\arguments{
\item{geneList}{order ranked geneList}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{pvalueCutoff}{p value Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{verbose}{print message or not}

\item{seed}{set seed inside the function to make result reproducible. FALSE by default.}

\item{USER_DATA}{annotation data}

\item{nproc}{If not equal to zero sets BPPARAM to use nproc workers (default = 1)}

\item{by}{one of 'fgsea' or 'DOSE'}
}
\value{
feaResult object
}
\description{
generic function for gene set enrichment analysis (drug targets score list as geneList,
GSEA method is modified to accept geneList with large portion of zeros)
}
