% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer2.R
\name{gseGO2}
\alias{gseGO2}
\title{gseGO2}
\usage{
gseGO2(geneList, ont = "BP", OrgDb, keyType = "ENTREZID",
  exponent = 1, nproc = 1, nPerm = 1000, minGSSize = 2,
  maxGSSize = 500, pvalueCutoff = 0.05, pAdjustMethod = "BH",
  verbose = TRUE)
}
\arguments{
\item{geneList}{scored ranked geneList}

\item{ont}{one of "BP", "MF", "CC" or "ALL"}

\item{OrgDb}{OrgDb}

\item{keyType}{keytype of gene}

\item{exponent}{weight of each step}

\item{nproc}{If not equal to zero, sets BPPARAM to use nproc workers (default = 1)}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}
}
\value{
feaResult object
}
\description{
Modified Gene Set Enrichment Analysis (GSEA) of Gene Ontology.
The GSEA method is modified to accept geneList with large portion of zeros
}
