% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer2.R
\name{gseKEGG2}
\alias{gseKEGG2}
\title{gseKEGG2}
\usage{
gseKEGG2(geneList, organism = "hsa", keyType = "kegg", exponent = 1,
  nproc = 1, nPerm = 1000, minGSSize = 10, maxGSSize = 500,
  pvalueCutoff = 0.05, pAdjustMethod = "BH", verbose = TRUE,
  use_internal_data = FALSE, seed = FALSE, by = "fgsea")
}
\arguments{
\item{geneList}{order ranked geneList}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{exponent}{weight of each step}

\item{nproc}{If not equal to zero, sets BPPARAM to use nproc workers (default = 1)}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{seed}{logical}

\item{by}{one of 'fgsea' or 'DOSE'}
}
\value{
feaResult object
}
\description{
Gene Set Enrichment Analysis of KEGG (drug targets score list as geneList)
}
\author{
Yu Guangchuang
}
