% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{lincs_expr_inst_info}
\alias{lincs_expr_inst_info}
\title{Instance Information of LINCS Expression Database}
\format{
A \code{tibble} object with 38,824 rows and 3 columns.
}
\usage{
lincs_expr_inst_info
}
\description{
It is a tibble of 3 columns containing compound treatment information of 
GEP instances in the LINCS expression database. 
The columns contain the compound name, cell type and perturbation type 
(all of them are compound treatment, trt_cp).
}
\examples{
# Load object
data(lincs_expr_inst_info)
head(lincs_expr_inst_info)
}
\keyword{datasets}
