% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{lincs_pert_info}
\alias{lincs_pert_info}
\title{LINCS 2017 Perturbation Information}
\format{
A \code{tibble} object with 8,140 rows and 40 columns.
}
\usage{
lincs_pert_info
}
\description{
It is a tibble containing annotation information of compounds in LINCS 2017 
database including perturbation name, type, whether in touchstone database, 
INCHI key, canonical smiles, PubChem CID as well as annotations from ChEMBL
database, including ChEMBL ID, DrugBank ID, max FDA phase, therapeutic flag,
first approval, indication class, mechanism of action, disease efficacy et al.
}
\examples{
# Load object
data(lincs_pert_info)
lincs_pert_info
}
\keyword{datasets}
