% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{lincs_pert_info2}
\alias{lincs_pert_info2}
\title{LINCS 2020 Perturbation Information}
\format{
A \code{tibble} object with 34419 rows and 48 columns.
}
\usage{
lincs_pert_info2
}
\description{
It is a tibble containing annotation information of compounds in LINCS 2020 
beta database including perturbation id, perturbation name, canonical smiles,
Inchi key, compound aliases, target and MOA. The PubChem CID and many other
annotations from ChEMBL database were obtained from 2017 LINCS pert info by
by left joining with pert_iname.
}
\examples{
# Load object
data(lincs_pert_info2)
lincs_pert_info2
}
\keyword{datasets}
