% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{lincs_sig_info}
\alias{lincs_sig_info}
\title{LINCS Signature Information}
\format{
A \code{tibble} object with 45,956 rows and 3 columns.
}
\usage{
lincs_sig_info
}
\description{
It is a tibble of 3 columns containing treatment information of GESs in the
LINCS database. The columns contain the perturbation name, cell
type and perturbation type (all of them are compound treatment, trt_cp).
}
\examples{
# Load object
data(lincs_sig_info)
head(lincs_sig_info)
}
\keyword{datasets}
