% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{lincsEnrich}
\alias{lincsEnrich}
\title{lincsEnrich}
\usage{
lincsEnrich(se, upset, downset, sortby = "NCS", type = 1,
  output = "all", ES_NULL = "Default", taurefList = "Default",
  minTauRefSize = 500, chunk_size = 5000)
}
\arguments{
\item{se}{SummarizedExperiment object}

\item{upset}{character vector represents up regulated gene sets}

\item{downset}{character vector represents down regulated gene sets}

\item{sortby}{by which the result data frame is sorted}

\item{type}{exponent of GSEA algorithm}

\item{output}{one of `esonly` and `all`.}

\item{ES_NULL}{path to the ES_NULL file. ES_null distribution is generated with random queryies for computing nominal P-values for ES by 
`randQueryES_slurm` or `randQueryES_local` function. If `ES_NULL` is set as `Default`, it uses the ES_null distribution that we generated.}

\item{taurefList}{path to the "taurefList.rds" file. If set as `Default`, it uses the "taurefList.rds" file that we generated.}

\item{minTauRefSize}{minimum size of reference data to compute Tau score}

\item{chunk_size}{size of chunk per processing}
}
\value{
data.frame
}
\description{
lincsEnrich
}
