% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list2df}
\alias{list2df}
\title{Named list to data frame}
\usage{
list2df(list, colnames)
}
\arguments{
\item{list}{input list with names slot}

\item{colnames}{character vector of length 2, indicating the column names
of the returned data.frame}
}
\value{
data.frame
}
\description{
Convert a list with names that have one to many mapping relationships to
a data.frame of two columns, one column is names, the other column is the
unlist elements
}
\examples{
list <- list("n1"=c("e1", "e2", "e4"), "n2"=c("e3", "e5"))
list2df(list, colnames=c("name", "element"))
}
