% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_rev}
\alias{list_rev}
\title{Reverse list}
\usage{
list_rev(list)
}
\arguments{
\item{list}{input list with names slot}
}
\value{
list
}
\description{
Reverse list from list names to elements mapping to elements to names mapping.
}
\examples{
list <- list("n1"=c("e1", "e2", "e4"), "n2"=c("e1", "e5"))
list_rev(list)
}
