% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabs_internal.R
\name{mabs_internal}
\alias{mabs_internal}
\title{mabs_internal}
\usage{
mabs_internal(geneList, nPerm, minGSSize, maxGSSize, pvalueCutoff,
  pAdjustMethod, USER_DATA)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{pvalueCutoff}{p value Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{USER_DATA}{annotation data}
}
\value{
mabsResult object
}
\description{
generic function for meanAbs analysis
}
\author{
Yuzhu Duan
}
