% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabsAnalyser.R
\name{mabsGO}
\alias{mabsGO}
\title{mabsGO}
\usage{
mabsGO(geneList, ont = "BP", OrgDb, keyType = "SYMBOL", nPerm = 1000,
  minGSSize = 10, maxGSSize = 500, pvalueCutoff = 0.05,
  pAdjustMethod = "BH")
}
\arguments{
\item{geneList}{order ranked geneList}

\item{ont}{one of "BP", "MF", "CC" or "ALL"}

\item{OrgDb}{OrgDb}

\item{keyType}{keytype of gene}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}
}
\value{
mabsResult object
}
\description{
MeanAbs Analysis of Gene Ontology
}
\author{
Yuzhu Duan
}
