% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_res_summary.R
\name{ncs_grp}
\alias{ncs_grp}
\title{ncs_grp}
\usage{
ncs_grp(tib, grp1, grp2)
}
\arguments{
\item{tib}{tibble object from result(gessResult), the 'gessResult' object is from 'gess_lincs' method}

\item{grp1}{character vector, group 1 of cell types, e.g., tumor cell types}

\item{grp2}{character vector, group 2 of cell types, e.g., normal cell types}
}
\value{
tibble
}
\description{
Append two columns (NCSgrp1, NCSgrp2) to gess_lincs result
}
