% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feaResult.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,feaResult,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{feaResult,ANY}(x, type = "bar", ...)
}
\arguments{
\item{x}{A \code{feaResult} instance}

\item{type}{one of dot, bar, cnet or enrichMap}

\item{...}{Additional argument list}
}
\value{
plot
}
\description{
plot method generics
}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
