% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{process_ids}
\alias{process_ids}
\title{Return a subset of requested GCTX row/colum ids
out of the universe of all ids}
\usage{
process_ids(ids, all_ids, type = "rid")
}
\arguments{
\item{ids}{vector of requested ids. If \code{NULL}, no
subsetting is performed}

\item{all_ids}{vector of universe of ids}

\item{type}{flag indicating the type of ids being processed}
}
\value{
a list with the following elements
 \code{ids}: a character vector of the processed ids
 \code{idx}: an integer list of their corresponding indices in \code{all_ids}
}
\description{
Return a subset of requested GCTX row/colum ids
out of the universe of all ids
}
\details{
This is a low-level helper function
  which most users will not need to access directly
}
\seealso{
Other GCTX parsing functions: \code{\link{append.dim}},
  \code{\link{fix.datatypes}}, \code{\link{parse_gctx}},
  \code{\link{read.gctx.ids}}, \code{\link{read.gctx.meta}}
}
\concept{GCTX parsing functions}
\keyword{internal}
