% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{qSig-class}
\alias{qSig-class}
\title{Class "qSig"}
\description{
This class represents the query signature for GESS analysis
}
\section{Slots}{

\describe{
\item{\code{qsig}}{When "gess_method" is "CMAP", "LINCS" or "Fisher", "sig" is a list of two elements, up and down regulated gene sets of entrez ids. 
When "gess_method" is "gCMAP" or "SP", "sig" is a list of one element, named numeric vector represents GEP from DE analysis or gene expression values.}

\item{\code{gess_method}}{one of "CMAP","LINCS","gCMAP","Fisher","SP"}

\item{\code{refdb}}{\code{SummarizedExperiment} object or a HDF5 backend \code{SummarizedExperiment} object loaded via 
`loadHDF5SummarizedExperiment` function from the HDF5 datasets saved in a directory. 
where all the assays are \code{HDF5Array} or \code{DelayedArray} objects consist of genome-wide differential expression profiles (GEPs)
 (e.g. log2 ratios or z-scores) from various drug treatments or genetic perturbations. 
It represents reference database that the query signature is searched against. Can be existing public databases (CMAP or LINCS)
or custom database.}
}}

\keyword{classes}
