% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/qSig-accessors.R
\docType{methods}
\name{qSig}
\alias{qSig}
\alias{qSig}
\alias{qSig,list,character,SummarizedExperiment-method}
\alias{qSig}
\alias{qSig,matrix,character,SummarizedExperiment-method}
\title{qSig method}
\usage{
qSig(qsig, gess_method, refdb)

\S4method{qSig}{list,character,SummarizedExperiment}(qsig, gess_method, refdb)

\S4method{qSig}{matrix,character,SummarizedExperiment}(qsig, gess_method, refdb)
}
\arguments{
\item{qsig}{A list contains gene sets or GEP}

\item{gess_method}{one of "CMAP","LINCS","gCMAP","Fisher","SP"}

\item{refdb}{\code{SummarizedExperiment} object or a HDF5 backend \code{SummarizedExperiment} object loaded via 
`loadHDF5SummarizedExperiment` function from the HDF5 datasets saved on disk. 
For \code{SummarizedExperiment} object, all the assays are \code{DelayedArray} objects consist of genome-wide differential expression profiles (GEPs)
 (e.g. log2 ratios or z-scores) from various drug treatments or genetic perturbations. 
It represents reference database that the query signature is searched against. 
Can be existing public databases (CMAP or LINCS) or custom database.

`CMAP` HDF5 backend database can be downloaded at \url{http://biocluster.ucr.edu/~yduan004/CMAP_db/cmap.tar.gz}, 
untar the file by "tar -xzvf file.tar.gz" command, then load the \code{SummarizedExperiment} object via 
`loadHDF5SummarizedExperiment` function. The "assays" slot of the loaded SummarizedExperiment object contains the logFC scores generated 
from differential expression (DE) analysis with `limma` of the original CEL files generated by Affymetrix Chips. 
The CMAP Affy data is stored at \url{https://portals.broadinstitute.org/cmap/}.
The code used for DE analysis is vailable at \url{http://girke.bioinformatics.ucr.edu/longevityTools/mydoc/mydoc_longevityTools_eDRUG_05.html}

`LINCS_de_sub` HDF5 backend database can be downloaded at \url{http://biocluster.ucr.edu/~yduan004/LINCS_db/lincs42_sub.tar.gz},
untar the file by "tar -xzvf file.tar.gz" command, then load the \code{SummarizedExperiment} object via 
`loadHDF5SummarizedExperiment` function.}
}
\value{
"qSig" object

"qSig" object
}
\description{
qSig generic

qSig method

qSig method
}
