% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{queryReferenceDB}
\alias{queryReferenceDB}
\title{Create Query Reference DB for Tau Score Computation of lincsEnrich}
\usage{
queryReferenceDB(se, Nup = 150, Ndown = 150, ES_NULL = "Default",
  dest_path, partition)
}
\arguments{
\item{se}{`SummarizedExperiment` object represents refernce database}

\item{Nup}{number of up-regulated genes subsetted for each signature in reference database `se` as query signature}

\item{Ndown}{number of down-regulated genes subsetted for each signature in reference database `se` as query signature}

\item{ES_NULL}{path to the ES_NULL file. ES_null distribution is generated with random queryies for computing nominal P-values for ES by 
`randQueryES_slurm` or `randQueryES_local` function. If `ES_NULL` is set as `Default`, it uses the ES_null distribution that we generated.}

\item{dest_path}{path to the destination file, including the file name "taurefList.rds"}

\item{partition}{partition node used to submit jobs on cluster}
}
\value{
creat "taurefList.rds" file
}
\description{
Create Query Reference DB for Tau Score Computation of lincsEnrich
}
