% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{randQueryES_local}
\alias{randQueryES_local}
\title{ES_null Distribution with Random Queryies for Computing Nominal P-values for ES}
\usage{
randQueryES_local(se, N_queries = 1000, dest_ES_NULL_path)
}
\arguments{
\item{se}{SummarizedExperiment object loaded from HDF5 backend LICNS GSE92742 database via `loadHDF5SummarizedExperiment(dbpath)` function
`dbpath` is the direcotry path to the `lincs42` database}

\item{N_queries}{number of random queries}

\item{dest_ES_NULL_path}{file path to the generated "ES_NULL.txt" file}
}
\value{
ES_NULL.txt file
}
\description{
ES_null Distribution with Random Queryies for Computing Nominal P-values for ES
}
