% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_cmap.R
\name{rankMatrix}
\alias{rankMatrix}
\title{rankMatrix}
\usage{
rankMatrix(x, decreasing = TRUE)
}
\arguments{
\item{x}{GEPs}

\item{decreasing}{rank decreasingly}
}
\description{
Rank transform signature database provided as matrix
}
\examples{
\dontrun{
 utils::download.file("http://biocluster.ucr.edu/~yduan004/CMAP_db/degList.rds", 
 "degList.rds", quiet = TRUE)
 logMA <- readRDS("degList.rds")$logFC
}
}
