% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{readHDF5chunk}
\alias{readHDF5chunk}
\title{Import HDF5 Data into SummarizedExperiment Object}
\usage{
readHDF5chunk(h5file, colindex = seq_len(10), colnames = NULL)
}
\arguments{
\item{h5file}{character(1), path to HDF5 file}

\item{colindex}{integer vector, position index of the matrix columns to be 
imported}

\item{colnames}{character vector, names of the columns of the matrix to be 
imported. If 'colnames' is set, 'colindex' will be ignored.}
}
\value{
\code{SummarizedExperiment} object
}
\description{
Imports user-definable subsets of matrix data from an HDF5 file into a 
\code{SummarizedExperiment} object. The
corresponding HDF5 file is expected to have three data components named
'assay', 'colnames' and 'rownames' containing the numeric values, column
names and row names of a matrix, respectively.
}
\examples{
gctx <- system.file("extdata", "test_sample_n2x12328.gctx", 
                    package="signatureSearch")
h5file <- tempfile(fileext=".h5")
gctx2h5(gctx, cid=1:2, 
        new_cid=c('sirolimus__MCF7__trt_cp', 'vorinostat__SKB__trt_cp'), 
        h5file=h5file, overwrite=TRUE)
se <- readHDF5chunk(h5file, colindex=1:2)
}
\seealso{
\code{SummarizedExperiment}
}
