% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/accessor.R
\docType{methods}
\name{result}
\alias{result}
\alias{result}
\alias{result,gessResult-method}
\title{result method}
\usage{
result(x)

\S4method{result}{gessResult}(x)
}
\arguments{
\item{x}{gessResult object}
}
\value{
tibble
}
\description{
result generic

get 'result' slot of gessResult object
}
