% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{targetList}
\alias{targetList}
\title{Target list used in examples}
\format{An object of class \code{numeric} of length 1000.}
\usage{
targetList
}
\description{
A named numeric vector with GENE SYMBOL as names. It is a subset of the 
first 1000 elements from 'targets' slot of 'mgsea_res' in the vignette. 
The scores represent the weights of targets/genes in the target set of the
selected top 10 drugs.
}
\keyword{datasets}
