% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tarReduce}
\alias{tarReduce}
\title{Show Reduced Targets}
\usage{
tarReduce(vec, Ntar = 5)
}
\arguments{
\item{vec}{character vector, each element composed by a list of targets 
symbols separated by '; '}

\item{Ntar}{integer, for each element in the vec, number of targets to show}
}
\value{
character vector after reducing
}
\description{
Reduce number of targets for each element of a character vector by 
replacting the targets that beyond Ntar to '...'.
}
\examples{
vec <- c("t1; t2; t3; t4; t5; t6", "t7; t8")
vec2 <- tarReduce(vec, Ntar=5)
}
