% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vec_char_redu}
\alias{vec_char_redu}
\title{Reduce Number of Character}
\usage{
vec_char_redu(vec, Nchar = 50)
}
\arguments{
\item{vec}{character vector to be reduced}

\item{Nchar}{integer, for each element in the vec, number of characters to 
remain}
}
\value{
character vector after reducing
}
\description{
Reduce number of characters for each element of a character vector by 
replacting the part that beyond Nchar (e.g. 50) character to '...'.
}
\examples{
vec <- c(strrep('a', 60), strrep('b', 30))
vec2 <- vec_char_redu(vec, Nchar=50)
}
