% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_custom_db.R
\name{build_custom_db}
\alias{build_custom_db}
\title{build_custom_db}
\usage{
build_custom_db(df, dest_path)
}
\arguments{
\item{df}{data.frame, represents genoime-wide GEPs (log2FC, z-scores, 
intensity values, etc.) of compound or genetic treatments. 
Rownames (Gene IDs) should be included. The colnames should be of 
`(drug)__(cell)__(factor)` format, e.g., `sirolimus__MCF7__trt_cp`.}

\item{dest_path}{directory path to store the reference database, which is 
the hdf5 backed \code{SummarizedExperiment} object}
}
\value{
hdf5 backed \code{SummarizedExperiment} object
}
\description{
Build custom reference signature database for GESS
}
