% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_fea_res.R
\name{comp_fea_res}
\alias{comp_fea_res}
\title{Plot Comparing FEA Results}
\usage{
comp_fea_res(table_list, rank_stat = "pvalue", Nshow = 20)
}
\arguments{
\item{table_list}{a named list of tables extracted from feaResults from 
different FEA methods. The names of the list is required, which could be 
name of the enrichment methods.}

\item{rank_stat}{character(1), column name of the enrichment statisic used 
for ranking the functional categories, e.g. 'pvalue' or 'p.adjust'. Note,
the column name should exist in each table in the 'table_list'.}

\item{Nshow}{integer, number of top re-ranking functional categories shown 
in the plot}
}
\value{
igraph object
}
\description{
Plot top re-ranking functional categories from different enrichment methods.
}
\details{
The `comp_fea_res` function re-ranks the functional categories from 
differentFEA methods by using the mean rank of each functional category
across the 5 FEA methods, the lower of the mean rank, the topper of the 
functional category. Since the functional categories have different number 
of support methods, the mean rank of a functional category is adjusted by 
multiplying the total number of methods and divided by the number of support 
methods. In this way, if a functional category is supported by only one 
method, the mean rank will be increased 5 times for penalty.
}
\examples{
data(drugs)
# table_list = list("dup_hyperG" = result(dup_hyperG_res), 
#                   "mGSEA" = result(mgsea_res), 
#                   "mabs" = result(mabs_res), 
#                   "hyperG" = result(hyperG_res), 
#                   "GSEA" = result(gsea_res))
# comp_fea_res(table_list, rank_stat="pvalue", Nshow=20)
}
