% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_fea_res.R
\name{comp_fea_res}
\alias{comp_fea_res}
\title{plot comparison of FEA enrichment results}
\usage{
comp_fea_res(dup_hyperG_res = NULL, mgsea_res = NULL,
  mabs_res = NULL, hyperG_res = NULL, gsea_res = NULL, Ntop = 20,
  type = "GO")
}
\arguments{
\item{dup_hyperG_res}{feaResult object from 'dup_hyperG' method}

\item{mgsea_res}{feaResult object from 'mGSEA' method}

\item{mabs_res}{feaResult object from 'mabs' method}

\item{hyperG_res}{feaResult object from 'hyperG' method}

\item{gsea_res}{feaResult object from 'GSEA' method}

\item{Ntop}{number of top enriched categories from each method used to do 
comparison. By default, the plot only compare the top 20 ranking terms}

\item{type}{"GO" or "KEGG", indicating the type of enriched terms}
}
\value{
igraph object
}
\description{
Plot consistency comparison of enriched terms from FEA methods
}
\details{
The enriched GO/KEGG categories were re-ranked by putting the terms 
supported by more methods to topper. For terms supported by the same number
of methods, the lower of the mean ranks of the terms across 
enrichment methods, the topper of the terms.
}
\examples{
\dontrun{
  comp_fea_res(dup_hyperG_res, mgsea_res, mabs_res, 
               hyperG_res, gsea_res, Ntop=20, type="GO")
}
}
