% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_fea_res.R
\name{comp_fea_res}
\alias{comp_fea_res}
\title{Plot for Comparing Ranking Results of FEA Methods}
\usage{
comp_fea_res(table_list, rank_stat = "pvalue", Nshow = 20)
}
\arguments{
\item{table_list}{a named list of tibbles extracted from feaResult objects,
e.g. generated with different FEA methods.}

\item{rank_stat}{character(1), column name of the enrichment statisic used 
for ranking the functional categories, e.g. 'pvalue' or 'p.adjust'. Note,
the chosen column name needs to be present in each tibble of 'table_list'.}

\item{Nshow}{integer defining the number of the top functional categories 
to display in the plot after re-ranking them across FEA methods}
}
\value{
ggplot2 graphics object
}
\description{
Dot plot for comparing the top ranking functional categories from different 
functional enrichment analysis (FEA) results. The functional categories are 
plotted in the order defined by their mean rank across the corresponding 
FEA results.
}
\details{
The `comp_fea_res` function computes the mean rank for each functional 
category across different FEA result instances and then re-ranks them based 
on that.Since the functional categories are not always present in all 
enrichment results, the mean rank of a functional category is corrected by 
an adjustment factor that is the number of enrichment result methods used 
divided by the number of occurences of a functional category. For instance, 
if a functional category is only present in the result of one method, its 
mean rank will be increased accordingly. Subsequently, the re-ranked 
functional categories are compared in a dot plot where the colors represent 
the values of the enrichment statistic chosen under the \code{rank_stat} 
argument.
}
\examples{
data(drugs)
\dontrun{
table_list <- list("dup_hyperG" = result(dup_hyperG_res), 
                  "mGSEA" = result(mgsea_res), 
                  "mabs" = result(mabs_res), 
                  "hyperG" = result(hyperG_res), 
                  "GSEA" = result(gsea_res))
comp_fea_res(table_list, rank_stat="pvalue", Nshow=20)
}
}
