% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feaResult.R
\docType{methods}
\name{dotplot}
\alias{dotplot}
\alias{dotplot,feaResult-method}
\alias{dotplot,feaResult,ANY-method}
\title{dotplot method}
\usage{
dotplot(object, ...)

\S4method{dotplot}{feaResult}(object, x = "geneRatio", colorBy = "p.adjust",
  showCategory = 10, split = NULL, font.size = 12, title = "")
}
\arguments{
\item{object}{an instance of feaResult}

\item{...}{additional parameter}

\item{x}{variable for x axis}

\item{colorBy}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{showCategory}{number of category}

\item{split}{separate result by 'category' variable}

\item{font.size}{font size}

\item{title}{plot title}
}
\value{
plot
}
\description{
dotplot

dotplot for feaResult
}
\author{
Guangchuang Yu

Guangchuang Yu
}
