% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{drugs}
\alias{drugs}
\title{Drugs used in examples}
\format{An object of class \code{character} of length 10.}
\usage{
drugs
}
\description{
A character vector containing top 10 drugs in the GESS result 
from gess_lincs method in the vignette
}
\keyword{datasets}
