% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsea_GSEA.R
\name{dsea_GSEA}
\alias{dsea_GSEA}
\title{dsea_GSEA}
\usage{
dsea_GSEA(drugList, type = "GO", ont = "BP", exponent = 1, nPerm = 1000,
  minGSSize = 10, maxGSSize = 500, pvalueCutoff = 0.05,
  pAdjustMethod = "BH")
}
\arguments{
\item{drugList}{ranked lists of all drugs in the GESS result. The scores of the corresponding GESS method can be used for ranking the drugs as it is 
required by the GSEA algorithm. The drugs with zero scores are excluded in the GESS results of CMAP, gCMAP and CMAP-LINCS methods}

\item{type}{one of "GO" or "KEGG"}

\item{ont}{one of "BP", "MF", "CC" or "GO"}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}
}
\value{
A \code{feaResult} instance.
}
\description{
Gene Set Enrichment Analysis of drug Ontology
}
\author{
Yuzhu Duan
}
