% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsea_GSEA.R
\name{dsea_GSEA}
\alias{dsea_GSEA}
\title{GSEA method for DSEA}
\usage{
dsea_GSEA(drugList, type = "GO", ont = "BP", exponent = 1,
  nPerm = 1000, minGSSize = 10, maxGSSize = 500,
  pvalueCutoff = 0.05, pAdjustMethod = "BH")
}
\arguments{
\item{drugList}{scored ranked list of all drugs in the GESS result. 
The similarity scores of the corresponding GESS method can be used for 
ranking the drugs as it is required by the GSEA algorithm. 
The drugs with zero scores are excluded.}

\item{type}{one of "GO", "KEGG" or "MOA"}

\item{ont}{one of "BP", "MF", "CC" or "ALL" if type is "GO"}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of drug sets annotated by ontology term 
after drug to functional category mappings. If type is "MOA", it is
recommended to set minGSSize as 2 since some MOA categories only contain 2 
drugs}

\item{maxGSSize}{maximal size of drug sets annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{p value adjustment method,
one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}
}
\value{
\code{\link{feaResult}} object, 
represents enriched functional categories.
}
\description{
The classical GSEA method is used to do enrichment analysis on a scored
ranked drug list after mapping drugs to functional categories via 
drug-target links in DrugBank, CLUE and STITCH databases for GO and KEGG 
pathway enrichment. It can also be used to get the enriched MOAs of a 
scored ranked drug list.
}
\details{
In the GESS/FEA workflow, the ranked drug list consists of all drugs in the 
GESS result. The similarity scores of the corresponding GESS method can be 
used for ranking the drugs as it is required by the GSEA algorithm. The drugs
with zero scores are excluded.
}
\examples{
db_path <- system.file("extdata", "sample_db.h5", 
                       package = "signatureSearch")
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_lincs <- qSig(query = list(upset=upset, downset=downset), 
                   gess_method = "LINCS", refdb = db_path)
lincs <- gess_lincs(qsig_lincs, sortby="NCS", tau=FALSE)
dl <- abs(result(lincs)$NCS); names(dl) <- result(lincs)$pert
dl <- dl[dl>0]
dl <- dl[!duplicated(names(dl))]
## GO annotation system
\dontrun{
gsea_res <- dsea_GSEA(drugList=dl, type="GO", ont="MF", exponent=1, 
                      nPerm=1000, pvalueCutoff=0.2, minGSSize=5)
                      result(gsea_res)
}
## KEGG annotation system
gsea_k_res <- dsea_GSEA(drugList=dl, type="KEGG", exponent=1, nPerm=1000, 
                        pvalueCutoff=0.5, minGSSize=5)
result(gsea_k_res)
}
\references{
GSEA algorithm: 
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., 
Gillette, M. A., … Mesirov, J. P. (2005). Gene set enrichment analysis: a 
knowledge-based approach for interpreting genome-wide expression profiles. 
Proceedings of the National Academy of Sciences of the United States of
America, 102(43), 15545–15550. \url{https://doi.org/10.1073/pnas.0506580102}
}
\seealso{
\code{\link{feaResult}}, \code{\link{fea}},
         \code{\link[signatureSearchData]{dtlink_db_clue_sti}}
}
