% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsea_hyperG.R
\name{dsea_hyperG}
\alias{dsea_hyperG}
\title{Drug GO Enrichment Analysis of drug sets.
Given a vector of drugs, this function will return the enriched GO categories with hypergeometric test after FDR control.}
\usage{
dsea_hyperG(drugs, type = "GO", ont = "BP", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", qvalueCutoff = 0.2, minGSSize = 10,
  maxGSSize = 500)
}
\arguments{
\item{drugs}{a vector of drug name.}

\item{type}{one of "GO" or "KEGG"}

\item{ont}{One of "MF", "BP", and "CC" or "ALL".}

\item{pvalueCutoff}{Cutoff value of p value.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{qvalueCutoff}{q value cutoff}

\item{minGSSize}{minimal size of drugs annotated by Ontology term for testing after drug-to-GO category mapping.}

\item{maxGSSize}{maximal size of drugs annotated for testing}
}
\value{
A \code{feaResult} instance.
}
\description{
Drug GO Enrichment Analysis of drug sets.
Given a vector of drugs, this function will return the enriched GO categories with hypergeometric test after FDR control.
}
\seealso{
\code{\link{feaResult-class}}
}
\author{
Yuzhu Duan
}
