% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtnetplot.R
\name{dtnetplot}
\alias{dtnetplot}
\title{dtnetplot}
\usage{
dtnetplot(drugs, set, ont = NULL, ...)
}
\arguments{
\item{drugs}{drugs used to build d-t network}

\item{set}{could be GO term ID, KEGG pathway ID or a character vector of gene set with gene SYMBOLs.}

\item{ont}{if the `set` is GO term ID, `ont` is the ontology that GO term is belong to. One of "BP", "MF", "CC" or "ALL"}

\item{...}{additional parameters}
}
\value{
visNetwork plot
}
\description{
dtnetplot
}
