% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMOA.R
\name{enrichMOA}
\alias{enrichMOA}
\title{MOA enrichment analysis via hypergeometric test}
\usage{
enrichMOA(gene, pvalueCutoff = 0.05, pAdjustMethod = "BH",
  qvalueCutoff = 0.2)
}
\arguments{
\item{gene}{a vector of gene entrez ids.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", 
"bonferroni", "BH", "BY", "fdr", "none"}

\item{qvalueCutoff}{qvalue cutoff}
}
\value{
A \code{feaResult} instance.
}
\description{
Given a vector of gene entrez ids, this function will return the enriched 
MOA categories after FDR control. The universe for the hypergeometric test
are all the genes in the MOA annotation obtained from ChEMBL database.
}
\examples{
library(org.Hs.eg.db)
data(geneList, package="DOSE")
\dontrun{
emoa <- enrichMOA(gene = names(geneList)[seq(3)])
head(emoa)
}
}
\seealso{
\code{\link{feaResult-class}}
}
