% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{feaResult-class}
\alias{feaResult-class}
\alias{feaResult}
\title{Class "feaResult"}
\description{
This class represents the result of functional enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame from FEA analysis}

\item{\code{organism}}{only "human" supported}

\item{\code{ontology}}{biological ontology}

\item{\code{drugs}}{Drug IDs}

\item{\code{targets}}{Target IDs of drugs in DrugBank/LINCS/STITCH databases or 
target list with scores.}
}}

\author{
Yuzhu Duan
}
\keyword{classes}
