% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{feaResult-class}
\alias{feaResult-class}
\alias{feaResult}
\title{Class "feaResult"}
\description{
This class represents the result of functional enrichment analysis.
}
\details{
Description of the common columns in the result table from different 
enrichment methods:
\itemize{
    \item ont: in case of GO, one of BP, MF, CC, or ALL
    \item ID: GO or KEGG IDs
    \item Description: description of functional category
    \item p.adjust: p-value adjusted for multiple hypothesis testing based 
    on method specified under pAdjustMethod argument
    \item qvalue: q value calculated with R’s qvalue function to control FDR
    \item itemID: IDs of items (genes for TSEA, drugs for DSEA) overlapping 
    among test and annotation sets.
    \item setSize: size of the functional category
}
}
\section{Slots}{

\describe{
\item{\code{result}}{tibble representing enriched functional categories}

\item{\code{organism}}{only "human" supported}

\item{\code{ontology}}{biological ontology}

\item{\code{drugs}}{Drug IDs}

\item{\code{targets}}{Target IDs of drugs in DrugBank/LINCS/STITCH databases or 
target list with scores.}
}}

\references{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., 
Gillette, M. A., … Mesirov, J. P. (2005). Gene set enrichment analysis: a 
knowledge-based approach for interpreting genome-wide expression profiles. 
Proceedings of the National Academy of Sciences of the United States of
America, 102(43), 15545–15550. \url{https://doi.org/10.1073/pnas.0506580102}
}
\author{
Yuzhu Duan
}
\keyword{classes}
