% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{feaResult-class}
\alias{feaResult-class}
\alias{feaResult}
\title{Class "feaResult"}
\description{
This class represents the result of functional enrichment analysis.
}
\details{
The description of the tibble columns in the result slot:
\itemize{
    \item ID: GO term or KEGG pathway ID.
    \item Description: description of the functional categories
    \item GeneRatio: ratio of genes in the test set that are annotated
    at a specific GO node or KEGG pathway
    \item BgRatio: ratio of background genes that are annotated
    at a specific GO node or KEGG pathway
    \item pvalue: p value of the enrichment
    \item p.adjust: p value adjusted for multiple hypothesis testing using
    'p.adjust' function with defined method. 
    \item qvalue: q value
    \item geneID: genes/drugs overlapped between test set and annotation 
    sets.
    \item setSize: size of the functional category
    \item enrichmentScore: Enrichment Score (ES) from the GSEA algorithm 
    (Subramanian et al., 2005). It represents the degree to which a set S 
    is over-represented at the top or bottom of the scored ranked list L. 
    The score is calculated by walking down the list L, 
    increasing a running-sum statistic when we encounter a gene in S and 
    decreasing when it is not. The magnitude of the increment depends on 
    the gene scores (e.g., correlation of the gene with phenotype). 
    The ES is the maximum deviation from zero encountered in the random 
    walk; it corresponds to a weighted Kolmogorov-Smirnov-like statistic.
    \item NES: normalized enrichment score. The positive and negative
    enrichment scores are normalized separately by permutating the 
    gene labels of the gene list L 'nPerm' times and dividing the 
    enrichment score by mean of the permutaion ES with the same sign.
    \item pvalue: The nominal p-value of the ES is calculated using 
    permutation test. Specifically, the gene labels of the gene list L were
    permuted and the ES of the gene set was recomputed for the permutated 
    data, which generate a null distribution for the ES. The p-value of the 
    observed ES is then calculated relative to this null distribution.
    \item p.adjust: p values adjusted for multiple hypothesis testing 
    \item qvalues: q value calculated for FDR control 
    \item leadingEdge: genes in the gene set S (functional category) that 
    appear in the ranked list L at, or before, the point where the running 
    sum reaches its maximum deviation from zero. Can be interpreted as the 
    core of a gene set that accounts for the enrichment signal.
    \item ledge_rank: ranks of genes in 'leadingEdge' at the gene list L.
    \item mabs: Given a scored ranked gene list L, mabs(S) represents
    the mean absolute scores of the genes in set S. 
    \item Nmabs: In order to adjust for size variations in gene set S, 
    'nPerm' random permutations of L are performed to determine 
    permutation mabs. Subsequently, mabs(S) is normalized by subtracting 
    the median of the permutation mabs and then dividing by its standard 
    deviation yielding the normalized scores Nmabs(S).
}
}
\section{Slots}{

\describe{
\item{\code{result}}{tibble representing enriched functional categories}

\item{\code{organism}}{only "human" supported}

\item{\code{ontology}}{biological ontology}

\item{\code{drugs}}{Drug IDs}

\item{\code{targets}}{Target IDs of drugs in DrugBank/LINCS/STITCH databases or 
target list with scores.}
}}

\references{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert, B. L., 
Gillette, M. A., … Mesirov, J. P. (2005). Gene set enrichment analysis: a 
knowledge-based approach for interpreting genome-wide expression profiles. 
Proceedings of the National Academy of Sciences of the United States of
America, 102(43), 15545–15550. \url{https://doi.org/10.1073/pnas.0506580102}
}
\author{
Yuzhu Duan
}
\keyword{classes}
