% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{feaResult-class}
\alias{feaResult-class}
\alias{feaResult}
\title{feaResult object}
\description{
The feaResult object stores the enrichment result table, organism 
information of the annotation system, and the ontology type of the GO 
annotation system. If the annotation system is KEGG, the latter will be 
'KEGG'. It also stores the input drugs used for the enrichment test, 
as well as their target information.
}
\section{Slots}{

\describe{
\item{\code{result}}{tibble object, this result table contains the
enriched functional categories (e.g. GO terms or KEGG pathways) ranked by 
the corresponding enrichment statistic. The result table can be extracted 
via \code{\link{result}} accessor function.

Description of the common columns in the result table from different 
enrichment methods:
\itemize{
    \item ont: in case of GO, one of BP, MF, CC, or ALL
    \item ID: GO or KEGG IDs
    \item Description: description of functional category
    \item p.adjust: p-value adjusted for multiple hypothesis testing based 
    on method specified under pAdjustMethod argument
    \item qvalue: q value calculated with R’s qvalue function to control FDR
    \item itemID: IDs of items (genes for TSEA, drugs for DSEA) overlapping 
    among test and annotation sets.
    \item setSize: size of the functional category
}}

\item{\code{organism}}{organism information of the annotation system, 
only 'human' supported.}

\item{\code{ontology}}{ontology type of the GO annotation system. If the annotation 
system is KEGG, it will be 'KEGG'}

\item{\code{drugs}}{input drugs used for the enrichment test}

\item{\code{targets}}{target information of the drugs in the defined drug-target 
annotation resource.}
}}

\keyword{classes}
