% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{feaResult-class}
\alias{feaResult-class}
\title{Class "feaResult"}
\description{
This class represents the result of functional enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame from FEA analysis}

\item{\code{organism}}{only "human" supported}

\item{\code{ontology}}{biological ontology}

\item{\code{drugs}}{Drug IDs}

\item{\code{targets}}{Target IDs of drugs in DrugBank/LINCS/STITCH databases or target list with scores.}

\item{\code{universe}}{background genes or drugs. For TSEA, it is all the genes in the corresponding annotation system (GO/KEGG). For DSEA, it is all the drugs
in the correspoinding annotation system (GO/KEGG) after drug-to-functional category mapping}

\item{\code{refSets}}{gene sets or drug sets in the corresponding annotation system}
}}

\author{
Yuzhu Duan
}
\keyword{classes}
