% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_cmap.R
\name{gess_cmap}
\alias{gess_cmap}
\title{CMAP method for GESS}
\usage{
gess_cmap(qSig, chunk_size = 5000)
}
\arguments{
\item{qSig}{\code{qSig} object, The 'gess_method' slot should be 'CMAP'.}

\item{chunk_size}{size of chunk per processing}
}
\value{
\code{gessResult} object, represents drugs in the reference database 
ranked by their similarity to the query signature
}
\description{
It uses query signature to search against the reference database in the 
\code{qSig} by CMAP method, which is an implementation of the original 
CMap method from Lamb et al, 2006.
}
\details{
Lamb et at., 2006 introduced the gene expression-based search method known as
Connectivity Map (CMap) where a GES database is searched with a query GES for
similar matches. It uses as query the most up- and down-regulated DEGs from 
a genome-wide expression experiment. The GES query is used to search a 
database of rank transformed GEPs and ranks the results by the degree of 
enrichment of the up- and down-regulated query genes on the top and bottom 
of the databases entries, respectively. The search results are a list of 
drugs that have similar or opposing GESs as the query. Similar GESs suggest 
similar physiological effects of the corresponding perturbagens. 

The CMAP method takes about 1.2 min on a single CPU core for querying with a 
single signature against 10,000 signatures in the LINCS database.

Description of the score columns in the gess_cmap tibble result:
\itemize{
    \item raw_score: bi-directional enrichment score (Kolmogorov-Smirnov 
    enrichment statistic) of up and down set in the query siganture
    \item scaled_score: raw_score was scaled to valules from 1 to -1 by 
    dividing the positive scores with maxmum positive score, and negative
    scores with minimum negative score.
}
}
\examples{
db_dir <- system.file("extdata", "sample_db", package = "signatureSearch")
sample_db <- loadHDF5SummarizedExperiment(db_dir)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_cmap <- qSig(qsig = list(upset=upset, downset=downset), 
                  gess_method = "CMAP", refdb = sample_db,
                  refdb_name="sample")
cmap <- gess_cmap(qSig=qsig_cmap, chunk_size=5000)
result(cmap)
}
\references{
For detailed description of the CMap method, please refer to: 
Lamb, J., Crawford, E. D., Peck, D., Modell, J. W., Blat, I. C., 
Wrobel, M. J., … Golub, T. R. (2006). The Connectivity Map: 
using gene-expression signatures to connect small molecules, genes, and 
disease. Science, 313(5795), 1929–1935. 
\url{https://doi.org/10.1126/science.1132939}
}
\seealso{
\code{\link{qSig}}, \code{\link{gessResult}}, \code{\link{gess}}
}
