% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_cor.R
\name{gess_cor}
\alias{gess_cor}
\title{gess_cor}
\usage{
gess_cor(qSig, method, chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot should be 'Cor'. The ‘qsig’ slot could be a matrix representing gene expression values, 
the reference database could also store genome-wide expression values of treatment samples}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. 
One of "spearman" (default), "kendall", or "pearson": can be abbreviated.}

\item{chunk_size}{size of chunk per processing}
}
\value{
gessResult object
}
\description{
Gene expression signature search with correlation coefficient
}
