% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_fisher.R
\name{gess_fisher}
\alias{gess_fisher}
\title{Fisher Search Method}
\usage{
gess_fisher(qSig, higher, lower, chunk_size = 5000)
}
\arguments{
\item{qSig}{\code{\link{qSig}} object defining the query signature including
the GESS method (should be 'Fisher') and the path to the reference
database. For details see help of \code{qSig} and \code{qSig-class}.}

\item{higher}{The 'higher' threshold. If not 'NULL', genes with a score 
larger than 'higher' will be included in the gene set with sign +1. 
At least one of 'lower' and 'higher' must be specified.}

\item{lower}{The lower threshold. If not 'NULL', genes with a score smaller 
than 'lower' will be included in the gene set with sign -1. 
At least one of 'lower' and 'higher' must be specified.}

\item{chunk_size}{number of database entries to process per iteration to 
limit memory usage of search.}
}
\value{
\code{\link{gessResult}} object, the result table contains the 
search results for each perturbagen in the reference database ranked by 
their signature similarity to the query.
}
\description{
In its iterative form, Fisher's exact test (Upton, 1992) can be used as Gene 
Expression Signature (GES) Search to scan GES databases for entries that 
are similar to a query GES.
}
\details{
When using the Fisher's exact test (Upton, 1992) as GES Search (GESS) method,
both the query and the database are composed of gene label sets, such as 
DEG sets.
}
\section{Column description}{

Descriptions of the columns specific to the Fisher method are given below. 
Note, the additional columns, those that are common among the GESS methods, 
are described in the help file of the \code{gessResult} object.

\itemize{
    \item pval: p-value of the Fisher's exact test.
    \item padj: p-value adjusted for multiple hypothesis testing using
    R's p.adjust function with the Benjamini & Hochberg (BH) method. 
    \item effect: z-score based on the standard normal distribution.
    \item LOR: Log Odds Ratio.
    \item nSet: number of genes in the GES in the reference
    database (gene sets) after setting the higher and lower cutoff.
    \item nFound: number of genes in the GESs of the reference
    database (gene sets) that are also present in the query GES.
    \item signed: whether gene sets in the reference database have signs, 
    representing up and down regulated genes when computing scores.  
}
}

\examples{
db_path <- system.file("extdata", "sample_db.h5", 
                       package = "signatureSearch")
#sample_db <- readHDF5chunk(db_path, colindex=1:100)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
#query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
#qsig_fisher <- qSig(query=query_mat, gess_method="Fisher", refdb=db_path)
#fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
#result(fisher)
}
\references{
Graham J. G. Upton. 1992. Fisher's Exact Test. J. R. Stat. Soc. Ser. A 
Stat. Soc. 155 (3). [Wiley, Royal Statistical Society]: 395-402. 
URL: http://www.jstor.org/stable/2982890
}
\seealso{
\code{\link{qSig}}, \code{\link{gessResult}}, \code{\link{gess}}
}
