% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_gcmap.R
\name{gess_gcmap}
\alias{gess_gcmap}
\title{gess_gcmap}
\usage{
gess_gcmap(qSig, higher, lower, add_bs_score = FALSE, chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot of 'qSig' should be 'gCMAP'}

\item{higher}{The 'higher' threshold. If not 'NULL', genes with a score larger than 'higher' will be included in the gene set with sign +1. 
At least one of 'lower' and 'higher' must be specified.}

\item{lower}{The lower threshold. If not 'NULL', genes with a score smaller than 'lower' will be included in the gene set with sign -1. 
At least one of 'lower' and 'higher' must be specified.}

\item{add_bs_score}{TRUE or FALSE. If true, bootstrap scores are added to measure the robustness of the result rankings.}

\item{chunk_size}{size of chunk per processing}
}
\value{
gessResult object represents a list of drugs in reference database ranked by their similarity to query signature
}
\description{
gCMAP method for GESS
}
