% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_gcmap.R
\name{gess_gcmap}
\alias{gess_gcmap}
\title{gCMAP Search Method}
\usage{
gess_gcmap(
  qSig,
  higher,
  lower,
  chunk_size = 5000,
  ref_trts = NULL,
  workers = 4
)
}
\arguments{
\item{qSig}{\code{\link{qSig}} object defining the query signature including
the GESS method (should be 'gCMAP') and the path to the reference database.
For details see help of \code{qSig} and \code{qSig-class}.}

\item{higher}{The 'upper' threshold. If not 'NULL', genes with a score 
larger than 'higher' will be included in the gene set with sign +1. 
At least one of 'lower' and 'higher' must be specified.}

\item{lower}{The lower threshold. If not 'NULL', genes with a score smaller 
than 'lower' will be included in the gene set with sign -1. 
At least one of 'lower' and 'higher' must be specified.}

\item{chunk_size}{number of database entries to process per iteration to 
limit memory usage of search.}

\item{ref_trts}{character vector. If users want to search against a subset 
of the reference database, they could set ref_trts as a character vector 
representing column names (treatments) of the subsetted refdb.}

\item{workers}{integer(1) number of workers for searching the reference
database parallelly, default is 4}
}
\value{
\code{\link{gessResult}} object, the result table contains the 
search results for each perturbagen in the reference database ranked by 
their signature similarity to the query.
}
\description{
Adapts the Gene Expression Signature Search (GESS) method from the gCMAP
package (Sandmann et al. 2014) to make it compatible with the database
containers and methods defined by \code{signatureSearch}. The specific GESS
method, called gCMAP, uses as query a rank transformed GES and the reference
database is composed of the labels of up and down regulated DEG sets.
}
\details{
The Bioconductor gCMAP (Sandmann et al. 2014) package provides access to a 
related but not identical implementation of the original CMAP algorithm 
proposed by Lamb et al. (2006). It uses as query a rank transformed GES and 
the reference database is composed of the labels of up and down regulated 
DEG sets. This is the opposite situation of the orignal CMAP method from 
Lamb et al (2006), where the query is composed of the labels of up and down 
regulated DEGs and the database contains rank transformed GESs.
}
\section{Column description}{

Descriptions of the columns specific to the gCMAP method are given below. 
Note, the additional columns, those that are common among the GESS methods, 
are described in the help file of the \code{gessResult} object.

\itemize{
    \item effect: Scaled bi-directional enrichment score corresponding to 
    the scaled_score under the CMAP result.
    \item nSet: Number of genes in the reference gene sets after applying
    the higher and lower cutoff.
    \item nFound: Number of genes in the reference gene sets that are 
    present in the query signature.
    \item signed: Whether the gene sets in the reference database have signs,
    e.g. representing up and down regulated genes when computing scores.
}
}

\examples{
db_path <- system.file("extdata", "sample_db.h5", 
                       package = "signatureSearch")
# library(SummarizedExperiment); library(HDF5Array)
# sample_db <- SummarizedExperiment(HDF5Array(db_path, name="assay"))
# rownames(sample_db) <- HDF5Array(db_path, name="rownames")
# colnames(sample_db) <- HDF5Array(db_path, name="colnames")
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
# query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
# qsig_gcmap <- qSig(query=query_mat, gess_method="gCMAP", refdb=db_path)
# gcmap <- gess_gcmap(qsig_gcmap, higher=1, lower=-1)
# result(gcmap)
}
\references{
Lamb, J., Crawford, E. D., Peck, D., Modell, J. W., Blat, I. C., 
Wrobel, M. J., Golub, T. R. (2006). The Connectivity Map: 
using gene-expression signatures to connect small molecules, genes, and 
disease. Science, 313 (5795), 1929-1935. 
URL: https://doi.org/10.1126/science.1132939

Sandmann, T., Kummerfeld, S. K., Gentleman, R., & Bourgon, R. 
(2014). gCMAP: user-friendly connectivity mapping with R. Bioinformatics , 
30 (1), 127-128. URL: https://doi.org/10.1093/bioinformatics/btt592
}
\seealso{
\code{\link{qSig}}, \code{\link{gessResult}}, \code{\link{gess}}
}
