% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{gess_lincs}
\alias{gess_lincs}
\title{gess_lincs}
\usage{
gess_lincs(qSig, ES_NULL = "Default", taurefList = "Default",
  sortby = "NCS", chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot should be 'LINCS'}

\item{ES_NULL}{path to the ES_NULL file. ES null distribution is generated with random queries for computing nominal P-values for ES by 
`randQueryES_slurm` or `randQueryES_local` function. 
If `ES_NULL` is set as `Default`, it uses the ES_null distribution that we generated against LINCS database.}

\item{taurefList}{path to the "taurefList.rds" file generated with `queryReferenceDB` function. 
If set as `Default`, it uses the "taurefList.rds" file that we generated against LINCS database.}

\item{sortby}{rank the GESS result by one of the following scores: `WTCS`, `NCS`, `Tau`, `NCSct` or `NA`}

\item{chunk_size}{size of chunk per processing}
}
\value{
data.frame of GESS result, a list of drugs in reference database ranked by their similarity to query signature
}
\description{
LINCS method for GESS
}
