% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{gess_lincs}
\alias{gess_lincs}
\title{LINCS method for GESS}
\usage{
gess_lincs(qSig, ES_NULL = "Default", tau = FALSE, sortby = "NCS",
  chunk_size = 5000)
}
\arguments{
\item{qSig}{`qSig` object, The 'gess_method' slot should be 'LINCS'}

\item{ES_NULL}{path to the ES_NULL file. ES null distribution is generated 
with random queries for computing nominal P-values for Enrichment Score in 
the gess result by `randQueryES` function. 
But it might take 3 days running 1000 random queries against LINCS
database locally without parallelizing jobs. 
To save time, it is recommended to parallelize with BiocParallel 
to run on CPU cores of a computer cluster with a scheduler (e.g. Slurm). For
codes used to generate the ES null distribution with Slurm scheduler on a 
computer cluster, please refer to `make-data.R` in `signatureSearch_data` 
package or at github \url{https://github.com/yduan004/signatureSearch_data} 

`ES_NULL` can also be set as `Default`, it uses the ES null 
distribution that we generated by running 1000 random queries against LINCS 
database.}

\item{tau, }{TRUE or FALSE, whether to compute the tau score. It is only 
meaningful when the "refdb" in "qSig" is LINCS since the reference
queries are from LINCS database.}

\item{sortby}{rank the GESS result by one of the following scores: 
`WTCS`, `NCS`, `Tau`, `NCSct` or `NA`}

\item{chunk_size}{size of chunk per processing}
}
\value{
\code{\link{gessResult}} object, containing drugs in the reference 
database ranked by their similarity to the query signature
}
\description{
It uses query signature to search against the reference database in the 
\code{qSig} by LINCS method, which is an implementation of the method from 
Subramanian et al, 2017.
}
\details{
Subramanian et al., 2017 introduced a new GESS method (\code{LINCS}) that is 
similar with the CMAP method but weights genes in the query set by using 
expression profiles with scores instead of rank transformed as reference 
database. Since now, no implementations of this algorithm available, it is 
implemented here. 

The LINCS method takes about 1.7 min on a single CPU core for querying 
with a single signature against 10,000 signatures in the LINCS database.
}
\examples{
db_dir <- system.file("extdata", "sample_db", package = "signatureSearch")
sample_db <- loadHDF5SummarizedExperiment(db_dir)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_lincs <- qSig(qsig = list(upset=upset, downset=downset), 
                   gess_method = "LINCS", refdb = sample_db,
                   refdb_name="sample")
lincs <- gess_lincs(qsig_lincs, sortby="NCS")
result(lincs)
}
\references{
For detailed description of the LINCS method, please refer to
Subramanian et al., 2017,
\url{https://www.sciencedirect.com/science/article/pii/S0092867417313090}
}
\seealso{
\code{\link{qSig}}, \code{\link{gessResult}}, \code{\link{gess}}
}
