% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_res_summary.R
\name{gess_res_vis}
\alias{gess_res_vis}
\title{GESS result visualization}
\usage{
gess_res_vis(gess_tb, drugs, col)
}
\arguments{
\item{gess_tb}{tibble in \code{\link{gessResult}} object, can be accessed
via \code{\link{result}} method.}

\item{drugs}{character vector, a list of interesting drugs}

\item{col}{name of the score column in 'gess_tb', e.g., "NCS"}
}
\value{
plot
}
\description{
The GESS result from a comprehensive signature database, e.g., LINCS,
contains similarity scores of drugs in different cell types, 
including normal and tumor. The similarity scores of a drug in cell types 
can be summarized and visualized. The summary method is according to the 
"Summarization Across Cell Lines" method from Subramanian et al., 2017.
}
\examples{
db_dir <- system.file("extdata", "sample_db", package = "signatureSearch")
sample_db <- loadHDF5SummarizedExperiment(db_dir)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_lincs <- qSig(qsig = list(upset=upset, downset=downset), 
                   gess_method = "LINCS", refdb = sample_db,
                   refdb_name="sample")
lincs <- gess_lincs(qsig_lincs, sortby="NCS")
data(drugs)
gess_res_vis(result(lincs), drugs, col="NCS")
}
\seealso{
Subramanian et al., 2017,
\url{https://www.sciencedirect.com/science/article/pii/S0092867417313090}
}
