% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_res_summary.R
\name{gess_res_vis}
\alias{gess_res_vis}
\title{GESS result visualization}
\usage{
gess_res_vis(gess_tb, drugs, col, cell_group = "all")
}
\arguments{
\item{gess_tb}{tibble in \code{\link{gessResult}} object, can be accessed
via \code{\link{result}} method.}

\item{drugs}{character vector, a list of interesting drugs}

\item{col}{name of the score column in 'gess_tb', e.g., "NCS". Can also be 
set as "rank", in this way, it will show the drug ranks in different cell
types.}

\item{cell_group}{one of "all", "normal", or "tumor". "all" represents both 
tumor and normal cells.}
}
\value{
plot
}
\description{
The GESS result from a comprehensive signature database, e.g., LINCS,
contains similarity scores of drugs in different cell types, 
including normal and tumor. The similarity scores of a drug in cell types 
can be summarized and visualized. The summary method is according to the 
"Summarization Across Cell Lines" method from Subramanian et al., 2017.
}
\examples{
db_path <- system.file("extdata", "sample_db.h5", 
                       package = "signatureSearch")
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_lincs <- qSig(query = list(upset=upset, downset=downset), 
                   gess_method = "LINCS", refdb = db_path)
lincs <- gess_lincs(qsig_lincs, sortby="NCS", tau=FALSE)
data(drugs)
gess_res_vis(result(lincs), drugs, col="NCS")
}
\references{
Subramanian, A., Narayan, R., Corsello, S. M., Peck, D. D., 
Natoli, T. E., Lu, X., … Golub, T. R. (2017). A Next Generation 
Connectivity Map: L1000 Platform and the First 1,000,000 Profiles. Cell, 
171(6), 1437–1452.e17. \url{https://doi.org/10.1016/j.cell.2017.10.049}
}
