% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gessResult-class}
\alias{gessResult-class}
\title{Class "gessResult"}
\description{
The class represents the result of GESS analysis
}
\section{Slots}{

\describe{
\item{\code{result}}{tibble from GESS analysis}

\item{\code{qsig}}{query signature}

\item{\code{gess_method}}{method for GESS analysis}

\item{\code{refdb}}{`SummarizedExperiment` object represents refrence signaure database}
}}

\keyword{classes}
