% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{gessResult-class}
\alias{gessResult-class}
\alias{gessResult}
\title{gessResult object}
\description{
The gessResult object stores the search result table, query signature, 
name of the GESS method and path to the reference database from the 
GESS methods.
}
\section{Slots}{

\describe{
\item{\code{result}}{tibble object, this result table contains the search results 
for each perturbagen in the reference database ranked by their signature 
similarity to the query. The result table can be extracted via 
\code{\link{result}} accessor function.

Description of the common columns from different GESS methods:
\itemize{
    \item pert: character, name of perturbagen (e.g. drug) in the reference 
    database
    \item cell: character, acronym of cell type
    \item type: character, perturbation type. In CMAP and LINCS 
    databases, the perturbation types are all compound treatment(trt_cp). 
    Users can build their custom signature database with other types of 
    perturbation, e.g., gene knockdown or overexpression via 
    \code{\link{build_custom_db}} function
    \item trend: character, up or down when reference signature is 
    positively or negatively connected with the query signature, 
    respectively.
    \item N_upset: integer, number of genes in the query up set
    \item N_downset: integer, number of genes in the query down set
    \item t_gn_sym: character, gene symbols of the corresponding drug 
    targets
}}

\item{\code{query}}{query signature}

\item{\code{gess_method}}{name of the GESS method}

\item{\code{refdb}}{path to the reference database}
}}

\keyword{classes}
