% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets.R
\name{get_targets}
\alias{get_targets}
\title{get drug targets in DrugBank/LINCS/STITCH}
\usage{
get_targets(drugs, database = "all")
}
\arguments{
\item{drugs}{character vector, a list of drug names}

\item{database}{one of "DrugBank", "LINCS", "STITCH" or "all"}
}
\value{
data.frame of drugs and target symbols
}
\description{
get targets in DrugBank/LINCS/STITCH for a given list of drugs
}
