% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets.R
\name{get_targets}
\alias{get_targets}
\title{get targets of query drugs}
\usage{
get_targets(drugs, database = "all")
}
\arguments{
\item{drugs}{a character vector storing a list of drug names}

\item{database}{one of "DrugBank", "LINCS", "STITCH" or "all"}
}
\value{
data.frame of drugs and target gene symbols
}
\description{
This function can be used to get protein targets of query drugs in 
DrugBank, CLUE and STITCH databases. A SQLite database storing drug-target
links in the above three databases can be found at 
\code{signatureSearch_data} package.
}
\examples{
data(drugs)
dt <- get_targets(drugs)
}
\seealso{
\code{dtlink_dt_clue_sti}
}
