% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets.R
\name{get_targets}
\alias{get_targets}
\title{Target Gene/Protein IDs for Query Drugs}
\usage{
get_targets(drugs, database = "all", verbose = TRUE, collapse = FALSE)
}
\arguments{
\item{drugs}{character vector of drug names}

\item{database}{drug-target annotation resource; one of 'DrugBank', 'CLUE',
'STITCH' or 'all'. If 'all', the targets from DrugBank, CLUE 
and STITCH databases will be combined.}

\item{verbose}{TRUE or FALSE, whether to print messages}

\item{collapse}{TRUE or FALSE. If setting as FALSE, the result is a 
data.frame of target gene symbols of each drug. If TRUE, the targets are 
collapsed to be a character vector with duplications if different drugs 
have the same targets.}
}
\value{
data.frame, one column contains the query drug names and the other 
target gene symbols if \code{collapse} is FALSE. A character vetor if
\code{collapse} is TRUE.
}
\description{
This function returns for a set of query drug names/ids the corresponding
target gene/protein ids. The required drug-target annotations are from
DrugBank, CLUE and STITCH. An SQLite database storing these drug-target
interactions based on the above three annotation resources is available in 
the \code{\link[signatureSearchData]{signatureSearchData}} package.
}
\examples{
data(drugs10)
dt <- get_targets(drugs10)
}
\seealso{
\code{\link[signatureSearchData]{dtlink_db_clue_sti}}
}
