% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_internal2.R
\name{GSEA_internal2}
\alias{GSEA_internal2}
\title{GSEA_internal2}
\usage{
GSEA_internal2(geneList, exponent, nPerm, minGSSize, maxGSSize,
  pvalueCutoff, pAdjustMethod, verbose, USER_DATA, nproc = 1)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{pvalueCutoff}{p value Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{verbose}{print message or not}

\item{USER_DATA}{annotation data}
}
\value{
feaResult object
}
\description{
Generic function for gene set enrichment analysis.
GSEA method is modified to accept geneList with large portion of zeros
}
