% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer2.R
\name{gseGO2}
\alias{gseGO2}
\title{modified GSEA method for GO enrichment analysis}
\usage{
gseGO2(geneList, ont = "BP", OrgDb, keyType = "SYMBOL", exponent = 1,
  nproc = 1, nPerm = 1000, minGSSize = 2, maxGSSize = 500,
  pvalueCutoff = 0.05, pAdjustMethod = "BH", verbose = TRUE)
}
\arguments{
\item{geneList}{numeric vector with gene SYMBOL as names, the genes are 
ranked decreasingly by their scores.}

\item{ont}{one of "BP", "MF", "CC" or "ALL"}

\item{OrgDb}{OrgDb, e.g., "org.Hs.eg.db".}

\item{keyType}{keytype of gene}

\item{exponent}{weight of each step}

\item{nproc}{If not equal to zero, sets BPPARAM to use nproc workers 
(default = 1)}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each gene set for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}
}
\value{
feaResult object
}
\description{
The modified Gene Set Enrichment Analysis (GSEA) of Gene Ontology supports
geneList with large portion of zeros and enrich GO terms at the top of the
geneList
}
\examples{
data(targetList)
library(org.Hs.eg.db)
gsego <- gseGO2(geneList=targetList, ont="MF", OrgDb=org.Hs.eg.db,
                pvalueCutoff = 1)
head(gsego)
}
