% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer2.R
\name{gseKEGG2}
\alias{gseKEGG2}
\title{modified GSEA method for KEGG enrichment analysis}
\usage{
gseKEGG2(geneList, organism = "hsa", keyType = "kegg", exponent = 1,
  nproc = 1, nPerm = 1000, minGSSize = 10, maxGSSize = 500,
  pvalueCutoff = 0.05, pAdjustMethod = "BH", verbose = TRUE)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{organism}{supported organism listed in '
\url{http://www.genome.jp/kegg/catalog/org_list.html}}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{exponent}{weight of each step}

\item{nproc}{If not equal to zero, 
sets BPPARAM to use nproc workers (default = 1)}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}
}
\value{
feaResult object
}
\description{
The modified GSEA algorithm supports geneList with large portion of zeros
and enrich KEGG pathways at the top of the geneList
}
\examples{
# Gene Entrez id should be used for KEGG enrichment
data(geneList, package="DOSE")
geneList[100:length(geneList)]=0
#gsekk <- gseKEGG2(geneList=geneList, pvalueCutoff = 1)
#head(gsekk)
}
