% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{lincs_pert_info}
\alias{lincs_pert_info}
\title{LINCS Perturbation Information}
\format{
A \code{tibble} object with 8,140 rows and 8 columns.
}
\usage{
lincs_pert_info
}
\description{
It is a tibble containing LINCS perturbation information including perturbation
name, type, whether in touchstone database, INCHI key, canonical smiles and 
PubChem CID.
}
\examples{
# Load object
data(lincs_pert_info)
lincs_pert_info
}
\keyword{datasets}
