% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabsAnalyser.R
\name{mabsKEGG}
\alias{mabsKEGG}
\title{mabsKEGG}
\usage{
mabsKEGG(geneList, organism = "hsa", keyType = "kegg", nPerm = 1000,
  minGSSize = 5, maxGSSize = 500, pvalueCutoff = 0.05,
  pAdjustMethod = "BH")
}
\arguments{
\item{geneList}{scored ranked geneList}

\item{organism}{supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}

\item{keyType}{one of "kegg", 'ncbi-geneid', 'ncib-proteinid' and 'uniprot'}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of genes annotated for testing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}
}
\value{
mabsResult object
}
\description{
meanAbs analysis of KEGG
}
