% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_conn.R
\name{moa_conn}
\alias{moa_conn}
\title{Summarize GESS results on MOA level}
\usage{
moa_conn(gess_tb, moa_cats = "default", cells = "normal")
}
\arguments{
\item{gess_tb}{table in \code{\link{gessResult}} object}

\item{moa_cats}{if set as "default", it uses MOA annotations from CLUE 
website \url{https://clue.io/}. 
Users can customize it as a `list` with names as MOA categories and 
contents as drug names.}

\item{cells}{one of "normal", "cancer" or "all", or a character vector 
containing selected cell types. It represents a group of interested 
cell types. 
\itemize{
  \item "all": all the 30 cell types in LINCS database;
  \item "normal": normal cell types in LINCS database as one group; 
  \item "tumor": tumor cell types in LINCS database as one group;
}}
}
\value{
data.frame
}
\description{
Summarize GESS results to get a list of MOA categories ranked by their 
connectivity to query signature in cells
}
\details{
Column description of the result table:

moa: Mechanism Of Action

cells: a group of interested cell types

mean_rank: mean ranks in GESS result of drugs in the MOA category

n_drug: number of drugs in the MOA category
}
\examples{
db_path <- system.file("extdata", "sample_db.h5", package = "signatureSearch")
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=seq_len(100))
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_fisher <- qSig(query=query_mat, gess_method="Fisher", refdb=db_path)
fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
res_moa <- moa_conn(result(fisher))
}
\seealso{
\code{\link{gessResult}}
}
