% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_conn.R
\name{moa_conn}
\alias{moa_conn}
\title{MOA connectivity}
\usage{
moa_conn(gess_tb, moa_cats = "default", cells = "normal")
}
\arguments{
\item{gess_tb}{table of GESS result}

\item{moa_cats}{if set as "default", it uses MOA annotaions in CLUE website (https://clue.io/). Users can customise it as a `list` object in R.}

\item{cells}{one of "normal", "cancer" or "all", or a character vector containing selected cell types. It represents a group of interested 
cell types. "all": all the 30 cell types in LINCS database; "normal": normal cell types in LINCS database as one group; 
"tumor": tumor cell types in LINCS database as one group;}
}
\value{
data.frame
}
\description{
Get MOA categories ranked by their connectivity to query signature in selected cells
}
