% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moa_conn.R
\name{moa_conn}
\alias{moa_conn}
\title{Summarize GESS Results on MOA Level}
\usage{
moa_conn(gess_tb, moa_cats = "default", cells = "normal")
}
\arguments{
\item{gess_tb}{tibble in \code{\link{gessResult}} object}

\item{moa_cats}{if set as "default", it uses MOA annotations from the CLUE 
website (https://clue.io). 
Users can customize it by providing a `list` of character vectors containing
drug names and MOA categories as list component names.}

\item{cells}{one of "normal", "cancer" or "all", or a character vector 
containing cell types of interest.
\itemize{
  \item "all": all cell types in LINCS database;
  \item "normal": normal cell types in LINCS database as one group; 
  \item "tumor": tumor cell types in LINCS database as one group;
}}
}
\value{
data.frame
}
\description{
Function summarizes GESS results on Mode of Action (MOA) level.
It returns a tabular representation of MOA 
categories ranked by their average signature search similarity to a query
signature.
}
\details{
Column description of the result table:

moa: Mechanism of Action (MOA)

cells: cell type information

mean_rank: mean rank of drugs in corresponding GESS result for each MOA 
category

n_drug: number of drugs in each MOA category
}
\examples{
db_path <- system.file("extdata", "sample_db.h5", package="signatureSearch")
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=seq_len(100))
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_fisher <- qSig(query=query_mat, gess_method="Fisher", refdb=db_path)
fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
res_moa <- moa_conn(result(fisher))
}
\seealso{
\code{\link{gessResult}}
}
