% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{qSig-class}
\alias{qSig-class}
\title{Class "qSig"}
\description{
This class stores the query signature, reference database
and GESS method used to search for similarity
}
\section{Slots}{

\describe{
\item{\code{query}}{When 'gess_method' is 'CMAP' or 'LINCS', 
it should be a list of two elements, which are up and down regulated 
gene sets of entrez ids.

When 'gess_method' is 'gCMAP', 'Fisher' or 'Cor', it should be a matrix 
representing gene expression profiles (GEPs) of treatment(s).}

\item{\code{gess_method}}{one of 'CMAP', 'LINCS', 'gCMAP', 'Fisher' or 'Cor'}

\item{\code{refdb}}{character(1), can be "cmap", "cmap_expr", "lincs", or 
"lincs_expr" if users want to use the existing CMAP/LINCS databases. 

If users want to use the custom signature database, 
it should be the file path to the HDF5 file generated with 
\code{\link{build_custom_db}} function or
generated from the source files of CMAP/LINCS databases according to 
the vignette in \code{\link[signatureSearchData]{signatureSearchData}}
package. The HDF5 file contains 
the reference signatures that the query signature is searched against.}

\item{\code{refdb_name}}{character(1), name of the reference database. Like "cmap",
"lincs" or other custom names.}
}}

\keyword{classes}
