% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/qSig-methods.R
\docType{methods}
\name{qSig}
\alias{qSig}
\alias{qSig,list,character,character,character-method}
\alias{qSig,matrix,character,character,character-method}
\title{qSig method}
\usage{
qSig(query, gess_method, refdb)

\S4method{qSig}{list,character,character}(query, gess_method, refdb)

\S4method{qSig}{matrix,character,character}(query, gess_method, refdb)
}
\arguments{
\item{gess_method}{one of 'CMAP', 'LINCS', 'gCMAP', 'Fisher' or 'Cor'}

\item{refdb}{character(1), can be "cmap", "cmap_expr", "lincs", or 
"lincs_expr" if users want to use the existing CMAP/LINCS databases. 

If users want to use the custom signature database, 
it should be the file path to the HDF5 file generated with 
\code{\link{build_custom_db}} function or
generated from the source files of CMAP/LINCS databases according to 
the vignette in \pkg{signatureSearchData} package. The HDF5 file contains 
the reference signatures that the query signature is searched against.}

\item{qsig}{When 'gess_method' is 'CMAP' or 'LINCS', 
it should be a list of two elements, which are up and down regulated gene 
sets of entrez ids.

When 'gess_method' is 'gCMAP', 'Fisher' or 'Cor', it should be a matrix 
representing gene expression profiles (GEPs) of treatment(s).}
}
\value{
\code{qSig} object
}
\description{
qSig generic

It builds a `qSig` object to store the query signature, reference database
and GESS method used to search for similarity
}
\examples{
db_path <- system.file("extdata", "sample_db.h5", 
package = "signatureSearch")
# Load sample_db as `SummarizedExperiment` object
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
# get "vorinostat__SKB__trt_cp" signature drawn from sample database
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
query = as.numeric(query_mat); names(query) = rownames(query_mat)
upset <- head(names(query[order(-query)]), 150)
downset <- tail(names(query[order(-query)]), 150)
qsig_lincs <- qSig(query = list(upset=upset, downset=downset), 
                   gess_method="LINCS", refdb=db_path)
qsig_gcmap <- qSig(qsig=query_mat, gess_method="gCMAP", refdb=db_path)
}
\seealso{
\code{\link{build_custom_db}}, 
\code{\link[signatureSearchData]{signatureSearchData}}
}
