% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/qSig-accessors.R
\docType{methods}
\name{qSig}
\alias{qSig}
\alias{qSig,list,character,SummarizedExperiment-method}
\alias{qSig,matrix,character,SummarizedExperiment-method}
\title{qSig method}
\usage{
qSig(qsig, gess_method, refdb)

\S4method{qSig}{list,character,SummarizedExperiment}(qsig, gess_method,
  refdb)

\S4method{qSig}{matrix,character,SummarizedExperiment}(qsig, gess_method,
  refdb)
}
\arguments{
\item{qsig}{When 'gess_method' is “CMAP” or “LINCS”, it should be a list of two elements, which are up and down regulated gene sets of entrez ids.
When 'gess_method' is “gCMAP”, “Fisher” or “Cor”, it should be a matrix representing gene expression profiles (GEPs) of treatment(s).}

\item{gess_method}{one of "CMAP", "LINCS", "gCMAP", "Fisher" or "Cor"}

\item{refdb}{\code{SummarizedExperiment} object, which can be HDF5 backend and loaded via 
`loadHDF5SummarizedExperiment` function from the HDF5 datasets saved on disk. 
For \code{SummarizedExperiment} object, The 'assays' slot should store a \code{DelayedArray} object consist of genome-wide gene expression 
profiles (GEPs) from various drug treatments or genetic perturbations. 
It represents reference database that the query signature is searched against. 
Can be existing public databases (CMAP or LINCS) or custom database.

`CMAP` HDF5 backend database can be downloaded at \url{http://biocluster.ucr.edu/~yduan004/CMAP_db/cmap.tar.gz}, 
untar the file by "tar -xzvf file.tar.gz" command, then load the \code{SummarizedExperiment} object via 
`loadHDF5SummarizedExperiment` function. The 'assays' slot of the loaded SummarizedExperiment object contains the logFC scores generated 
from differential expression (DE) analysis with `limma` of the original CEL files generated by Affymetrix Chips. 
The CMAP Affy data is stored at \url{https://portals.broadinstitute.org/cmap/}.
The code used for DE analysis is vailable at \url{http://girke.bioinformatics.ucr.edu/longevityTools/mydoc/mydoc_longevityTools_eDRUG_05.html}

`LINCS` HDF5 backend database can be downloaded at \url{http://biocluster.ucr.edu/~yduan004/LINCS_db/lincs42.tar.gz},
untar the file by "tar -xzvf file.tar.gz" command, then load the \code{SummarizedExperiment} object via 
`loadHDF5SummarizedExperiment` function.}
}
\value{
"qSig" object

"qSig" object
}
\description{
qSig generic
}
