% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{randQueryES}
\alias{randQueryES}
\title{Generate WTCS null distribution with random queries}
\usage{
randQueryES(h5file, N_queries = 1000, dest)
}
\arguments{
\item{h5file}{character(1), path to the HDF5 file representting the
reference database}

\item{N_queries}{number of random queries}

\item{dest}{file path to the generated "ES_NULL.txt" file}
}
\value{
write 'ES_NULL.txt' file to the 'dest'
}
\description{
Generate null distribution of WTCS with specified number of random queries 
against the reference database for computing nominal P-values for WTCS
in the `gess_lincs` result.
}
\examples{
db_path = system.file("extdata", "sample_db.h5", package="signatureSearch")
library(signatureSearchData)
randQueryES(h5file=db_path, N_queries=5, dest="ES_NULL.txt")
unlink("ES_NULL.txt")
}
\references{
Subramanian, A., Narayan, R., Corsello, S. M., Peck, D. D., Natoli, T. E., 
Lu, X., … Golub, T. R. (2017). A Next Generation Connectivity Map: L1000 
Platform and the First 1,000,000 Profiles. Cell, 171(6), 1437–1452.e17. 
\url{https://doi.org/10.1016/j.cell.2017.10.049}
}
\seealso{
\code{\link{gess_lincs}}
}
