% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_lincs.R
\name{randQueryES}
\alias{randQueryES}
\title{Generate ES null distribution with random queries}
\usage{
randQueryES(se, N_queries = 1000, dest_ES_NULL_path)
}
\arguments{
\item{se}{SummarizedExperiment object represents reference database}

\item{N_queries}{number of random queries}

\item{dest_ES_NULL_path}{file path to the generated "ES_NULL.txt" file}
}
\value{
ES_NULL.txt file
}
\description{
Generate ES null distribution with specified number of random queryies 
against the reference database (se) for computing nominal P-values for ES
in the `gess_lincs` result.
}
\examples{
dbpath = system.file("extdata", "sample_db", package="signatureSearch")
se = HDF5Array::loadHDF5SummarizedExperiment(dbpath)
randQueryES(se=se, N_queries=5, dest_ES_NULL_path="ES_NULL.txt")
unlink("ES_NULL.txt")
}
\references{
Subramanian et al., 2017,
\url{https://www.sciencedirect.com/science/article/pii/S0092867417313090}
}
\seealso{
\code{\link{gess_lincs}}
}
