% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feaResult-methods.R
\docType{methods}
\name{result}
\alias{result}
\alias{result,gessResult-method}
\alias{result,feaResult-method}
\title{result method}
\usage{
result(x)

\S4method{result}{gessResult}(x)

\S4method{result}{feaResult}(x)
}
\arguments{
\item{x}{\code{gessResult} or \code{feaResult} object}
}
\value{
tibble
}
\description{
result generic

get 'result' slot of gessResult object

get 'result' slot of feaResult object
}
