% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feaResult-methods.R
\docType{methods}
\name{result}
\alias{result}
\alias{result,gessResult-method}
\alias{result,feaResult-method}
\title{result method}
\usage{
result(x)

\S4method{result}{gessResult}(x)

\S4method{result}{feaResult}(x)
}
\arguments{
\item{x}{\code{gessResult} or \code{feaResult} object}
}
\value{
tibble
}
\description{
result generic

get 'result' slot of gessResult object

get 'result' slot of feaResult object
}
\examples{
db_dir <- system.file("extdata", "sample_db", package = "signatureSearch")
sample_db <- loadHDF5SummarizedExperiment(db_dir)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_fisher <- qSig(qsig=query_mat, gess_method="Fisher", refdb=sample_db,
                    refdb_name="sample")
fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
result(fisher)
data(drugs)
dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
                                  type = "GO", ont="MF")
result(dup_hyperG_res) 
}
