% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/feaResult-methods.R,
%   R/gessResult-methods.R
\docType{methods}
\name{result}
\alias{result}
\alias{result,feaResult-method}
\alias{result,gessResult-method}
\title{result method}
\usage{
result(x)

\S4method{result}{feaResult}(x)

\S4method{result}{gessResult}(x)
}
\arguments{
\item{x}{\code{gessResult} or \code{feaResult} object}
}
\value{
tibble
}
\description{
result generic

get 'result' slot of feaResult object

get 'result' slot of gessResult object
}
\examples{
data(drugs)
#dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
#                                 type = "GO", ont="MF")
#result(dup_hyperG_res) 
db_path <- system.file("extdata", "sample_db.h5", 
package = "signatureSearch")
# Load sample_db as `SummarizedExperiment` object
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
# get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_fisher <- qSig(query=query_mat, gess_method="Fisher", refdb=db_path)
fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
result(fisher)
}
