% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feaResult-methods.R, R/qSig-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gessResult-method}
\alias{show,feaResult-method}
\alias{show,qSig-method}
\title{show method}
\usage{
show(object)

\S4method{show}{feaResult}(object)

\S4method{show}{qSig}(object)
}
\arguments{
\item{object}{object used for show}
}
\value{
message
}
\description{
show qSig, gessResult, feaResult objects
}
\examples{
db_dir <- system.file("extdata", "sample_db", package = "signatureSearch")
sample_db <- loadHDF5SummarizedExperiment(db_dir)
## get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_fisher <- qSig(qsig=query_mat, gess_method="Fisher", refdb=sample_db,
                    refdb_name="sample")
fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
fisher
data(drugs)
dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
                                  type = "GO", ont="MF")
dup_hyperG_res 
}
