% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feaResult.R, R/qSig-accessors.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,gessResult-method}
\alias{show,feaResult-method}
\alias{show,qSig-method}
\title{show method}
\usage{
show(object)

\S4method{show}{feaResult}(object)

\S4method{show}{qSig}(object)
}
\arguments{
\item{object}{object used for show}
}
\value{
message
}
\description{
show qSig, gessResult, feaResult objects
}
