% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feaResult-methods.R, R/gessResult-methods.R,
%   R/qSig-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,feaResult-method}
\alias{show,gessResult-method}
\alias{show,qSig-method}
\title{show method}
\usage{
\S4method{show}{feaResult}(object)

show(object)

\S4method{show}{qSig}(object)
}
\arguments{
\item{object}{object used for show}
}
\value{
message
}
\description{
show qSig, gessResult, feaResult objects
}
\examples{
data(drugs)
# dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
#                                 type = "GO", ont="MF")
# dup_hyperG_res 
db_path <- system.file("extdata", "sample_db.h5", 
package = "signatureSearch")
# Load sample_db as `SummarizedExperiment` object
library(signatureSearchData)
sample_db <- readHDF5chunk(db_path, colindex=1:100)
# get "vorinostat__SKB__trt_cp" signature drawn from sample databass
query_mat <- as.matrix(assay(sample_db[,"vorinostat__SKB__trt_cp"]))
qsig_fisher <- qSig(query=query_mat, gess_method="Fisher", refdb=db_path,
                    refdb_name="sample")
fisher <- gess_fisher(qSig=qsig_fisher, higher=1, lower=-1)
fisher
}
