% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gess_res_summary.R
\name{sim_score_grp}
\alias{sim_score_grp}
\title{sim_score_grp}
\usage{
sim_score_grp(tib, grp1, grp2, score_column)
}
\arguments{
\item{tib}{tibble object from result(gessResult)}

\item{grp1}{character vector, group 1 of cell types, e.g., tumor cell types}

\item{grp2}{character vector, group 2 of cell types, e.g., normal cell types}

\item{score_column}{character, column name of similairity scores to be grouped}
}
\value{
tibble
}
\description{
Append two columns (score_column_grp1, score_column_grp2) to GESS result
}
