% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureSearch-package.r
\docType{data}
\name{targetList}
\alias{targetList}
\title{Target Sample Data Set}
\format{An object of class \code{numeric} of length 1000.}
\usage{
targetList
}
\description{
A named numeric vector with Gene Symbols as names. It is the first 1000
elements from the 'targets' slot of the 'mgsea_res' result object introduced 
in the vignette of this package. The scores represent the weights of the 
target genes/proteins in the target set of the selected top 10 drugs.
}
\examples{
# Load object
data(targetList)
targetList
}
\keyword{datasets}
