% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_dup_hyperG.R
\name{tsea_dup_hyperG}
\alias{tsea_dup_hyperG}
\title{Target Set Enrichment Analysis (TSEA) with Hypergeometric Test}
\usage{
tsea_dup_hyperG(drugs, universe = "Default", type = "GO", ont = "MF",
  pAdjustMethod = "BH", pvalueCutoff = 0.05, qvalueCutoff = 0.05,
  minGSSize = 5, maxGSSize = 500, dt_anno = "all")
}
\arguments{
\item{drugs}{character vector containing drug identifiers used for functional
enrichment testing. This can be the top ranking drugs from a GESS result. 
Internally, drug test sets are translated to the corresponding target protein
test sets based on the drug-target annotations provided under the 
\code{dt_anno} argument.}

\item{universe}{character vector defining the universe of genes/proteins. If
set as 'Default', it uses all genes/proteins present in the corresponding
annotation system (e.g. GO or KEGG). If 'type' is 'GO', it can be assigned
a custom vector of gene SYMBOL IDs. If 'type' is 'KEGG', the vector needs
to contain Entrez gene IDs.}

\item{type}{one of `GO` or `KEGG`}

\item{ont}{character(1). If type is `GO`, assign \code{ont} (ontology) one of
`BP`,`MF`, `CC` or `ALL`. If type is 'KEGG', \code{ont} is ignored.}

\item{pAdjustMethod}{p-value adjustment method, 
one of 'holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', 'fdr'}

\item{pvalueCutoff}{double, p-value cutoff}

\item{qvalueCutoff}{double, qvalue cutoff}

\item{minGSSize}{integer, minimum size of each gene set in annotation system}

\item{maxGSSize}{integer, maximum size of each gene set in annotation system}

\item{dt_anno}{drug-target annotation source. Currently, one of 'DrugBank',
'CLUE', 'STITCH' or 'all'. If 'dt_anno' is 'all', the targets from the
DrugBank, CLUE and STITCH databases will be combined. Usually, it is
recommended to set the 'dt_anno' to 'all' since it provides the most
complete drug-target annotations. Choosing a single
annotation source results in sparser drug-target annotations
(particularly CLUE), and thus less complete enrichment results.}
}
\value{
\code{\link{feaResult}} object, the result table contains the
enriched functional categories (e.g. GO terms or KEGG pathways) ranked by 
the corresponding enrichment statistic.
}
\description{
The \code{tsea_dup_hyperG} function performs Target Set Enrichment Analysis
(TSEA) based on a modified hypergeomtric test that supports test sets with
duplications. This is achieved by maintaining the frequency information of
duplicated items in form of weighting values.
}
\details{
The classical hypergeometric test assumes uniqueness in its test sets. To
maintain the duplication information in the test sets used for TSEA, the 
values of the total number of genes/proteins in the test set and the number 
of genes/proteins in the test set annotated at a functional category are
adjusted by maintating their frequency information in the test set rather 
than counting each entry only once. Removing duplications in TSEA would be 
inappropriate since it would erase one of the most important pieces of 
information of this approach.
}
\section{Column description}{

The TSEA results (including \code{tsea_dup_hyperG}) stored in the
\code{feaResult} object can be returned with the \code{result} method in
tabular format, here \code{tibble}. The columns of this \code{tibble} are
described below.
\itemize{
    \item GeneRatio: ratio of genes in the test set that are annotated at a 
    specific GO node or KEGG pathway
    \item BgRatio: ratio of background genes that are annotated
    at a specific GO node or KEGG pathway
    \item pvalue: raw p-value of enrichment test
}
Additional columns are described under the 'result' slot of the
\code{\link{feaResult}} object.
}

\examples{
data(drugs)
\dontrun{
## GO annotation system
dup_hyperG_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
                                  type = "GO", ont="MF", pvalueCutoff=0.05,
                                  pAdjustMethod="BH", qvalueCutoff = 0.1, 
                                  minGSSize = 10, maxGSSize = 500)
result(dup_hyperG_res)
}
## KEGG annotation system
dup_hyperG_k_res <- tsea_dup_hyperG(drugs = drugs, universe = "Default", 
                                    type = "KEGG", pvalueCutoff=0.1, 
                                    pAdjustMethod="BH", qvalueCutoff = 0.2, 
                                    minGSSize = 10, maxGSSize = 500)
result(dup_hyperG_k_res)
}
\seealso{
\code{\link{feaResult}}, \code{\link{fea}}
}
