% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_dup_hyperG.R
\name{tsea_dup_hyperG}
\alias{tsea_dup_hyperG}
\title{tsea_dup_hyperG}
\usage{
tsea_dup_hyperG(drugs, universe = "Default", type = "GO", ont = "MF",
  pAdjustMethod = "BH", pvalueCutoff = 0.05, qvalueCutoff = 0.05,
  minGSSize = 5, maxGSSize = 500)
}
\arguments{
\item{drugs}{query drug set used to do target set enrichment analysis (TSEA).
Can be top ranking drugs in GESS result.}

\item{universe}{background genes/targets. If set as `Default`, 
it represents all the annotated genes/targets in
the corresponding annotation system (e.g. GO or KEGG). 
If "type" is "GO", universe should be gene SYMBOL ids. If "type" is "KEGG",
universe should be gene entrez ids.}

\item{type}{one of `GO` or `KEGG`}

\item{ont}{if type is `GO`, set ontology. One of `BP`,`MF`,`CC` or `ALL`}

\item{pAdjustMethod}{p value adjustment method for p values 
in the enrichment result}

\item{pvalueCutoff}{p value Cutoff}

\item{qvalueCutoff}{qvalue Cutoff}

\item{minGSSize}{minimum size of each gene set in annotation system}

\item{maxGSSize}{maximum size of each gene set in annotation system}
}
\value{
feaResult object
}
\description{
drug targets GO/KEGG enrichment analysis by using `dup_hyperG` method
}
\author{
Yuzhu Duan (yduan004@ucr.edu)
}
