% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_mabs.R
\name{tsea_mabs}
\alias{tsea_mabs}
\title{Target Set Enrichment Analysis (TSEA) with meanAbs}
\usage{
tsea_mabs(drugs, type = "GO", ont = "MF", nPerm = 1000,
  pAdjustMethod = "BH", pvalueCutoff = 0.05, minGSSize = 5,
  maxGSSize = 500, dt_anno = "all")
}
\arguments{
\item{drugs}{character vector containing drug identifiers used for functional
enrichment testing. This can be the top ranking drugs from a GESS result. 
Internally, drug test sets are translated to the corresponding target protein
test sets based on the drug-target annotations provided under the 
\code{dt_anno} argument.}

\item{type}{one of `GO` or `KEGG`}

\item{ont}{character(1). If type is `GO`, assign \code{ont} (ontology) one of
`BP`,`MF`, `CC` or `ALL`. If type is 'KEGG', \code{ont} is ignored.}

\item{nPerm}{integer, permutation number used to calculate p-values}

\item{pAdjustMethod}{p-value adjustment method, 
one of 'holm', 'hochberg', 'hommel', 'bonferroni', 'BH', 'BY', 'fdr'}

\item{pvalueCutoff}{double, p-value cutoff}

\item{minGSSize}{integer, minimum size of each gene set in annotation system}

\item{maxGSSize}{integer, maximum size of each gene set in annotation system}

\item{dt_anno}{drug-target annotation source. Currently, one of 'DrugBank',
'CLUE', 'STITCH' or 'all'. If 'dt_anno' is 'all', the targets from the
DrugBank, CLUE and STITCH databases will be combined. Usually, it is
recommended to set the 'dt_anno' to 'all' since it provides the most
complete drug-target annotations. Choosing a single
annotation source results in sparser drug-target annotations 
(particularly CLUE), and thus less complete enrichment results.}
}
\value{
\code{\link{feaResult}} object, the result table contains the
enriched functional categories (e.g. GO terms or KEGG pathways) ranked by 
the corresponding enrichment statistic.
}
\description{
The meanAbs (mabs) method is a simple but effective functional enrichment
statistic (Fang et al., 2012). As required for TSEA, it supports query label
sets (here for target proteins/genes) with duplications by transforming them 
to score ranked label lists and then calculating mean absolute scores of 
labels in label set \eqn{S}.
}
\details{
The input for the mabs method is \eqn{L_tar}, the same as for mGSEA. In this
enrichment statistic, \eqn{mabs(S)}, of a label (e.g. gene/protein) set
\eqn{S} is calculated as mean absolute scores of the labels in \eqn{S}. In
order to adjust for size variations in label set \eqn{S}, 1000 random
permutations of \eqn{L_tar} are performed to determine \eqn{mabs(S,pi)}.
Subsequently, \eqn{mabs(S)} is normalized by subtracting the median of the
\eqn{mabs(S,pi)} and then dividing by the standard deviation of
\eqn{mabs(S,pi)} yielding the normalized scores \eqn{Nmabs(S)}. Finally, the
portion of \eqn{mabs(S,pi)} that is greater than \eqn{mabs(S)} is used as
nominal p-value (Fang et al., 2012). The resulting nominal p-values are
adjusted for multiple hypothesis testing using the Benjamini-Hochberg method.
}
\section{Column description}{

The TSEA results (including \code{tsea_mabs}) stored in the
\code{feaResult} object can be returned with the \code{result} method in
tabular format, here \code{tibble}. The columns in this \code{tibble}
specific to the \code{mabs} method are described below.
\itemize{
    \item mabs: given a scored ranked gene list \eqn{L}, \eqn{mabs(S)}
    represents the mean absolute scores of the genes in set \eqn{S}.
    \item Nmabs: \eqn{mabs(S)} normalized
}
Additional columns are described under the 'result' slot of the
\code{\link{feaResult}} object.
}

\examples{
data(drugs)
## GO annotation system
#mabs_res <- tsea_mabs(drugs=drugs, type="GO", ont="MF", nPerm=1000, 
#                      pvalueCutoff=0.05, minGSSize=5)
#result(mabs_res)
## KEGG annotation system
#mabs_k_res <- tsea_mabs(drugs=drugs, type="KEGG", nPerm=1000, 
#                        pvalueCutoff=0.05, minGSSize=5)
#result(mabs_k_res)
}
\references{
Fang, Z., Tian, W., & Ji, H. (2012). A network-based 
gene-weighting approach for pathway analysis. Cell Research, 22(3), 
565-580. URL: https://doi.org/10.1038/cr.2011.149
}
\seealso{
\code{\link{feaResult}}, \code{\link{fea}}, \code{\link{tsea_mGSEA}}
}
