% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsea_mGSEA.R
\name{tsea_mGSEA}
\alias{tsea_mGSEA}
\title{tsea_mGSEA}
\usage{
tsea_mGSEA(drugs, type = "GO", ont = "MF", nPerm = 1000,
  exponent = 1, pAdjustMethod = "BH", pvalueCutoff = 0.05,
  minGSSize = 2, maxGSSize = 500)
}
\arguments{
\item{drugs}{query drug set used to do TSEA. Can be top ranking drugs in GESS result.}

\item{type}{onoe of `GO` or `KEGG`}

\item{ont}{if type is `GO`, set ontology, one of `BP`,`MF`,`CC` or `ALL`}

\item{nPerm}{permutation numbers used to calculate p value}

\item{exponent}{weight of each step}

\item{pAdjustMethod}{p value adjustment method for p values in the enrichment result}

\item{pvalueCutoff}{p value cutoff}

\item{minGSSize}{minimum size of each gene set in annotation system}

\item{maxGSSize}{maximum size of each gene set in annotation system}
}
\value{
A \code{feaResult} instance
}
\description{
drug targets GO/KEGG enrichment analysis by using `mGSEA` method
}
\author{
Yuzhu Duan (yduan004@ucr.edu)
}
